% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atomic_plots_common_arguments}
\alias{atomic_plots_common_arguments}
\title{Common arguments for atomic plots}
\arguments{
\item{facet}{A logical value indicating whether to facet the plot.}

\item{facet_by}{A character string specifying the column name of the data frame to facet the plot.}

\item{facet_scales}{A character string specifying the scales of the facets.}

\item{nrow}{A numeric value specifying the number of rows in the facet.}

\item{ncol}{A numeric value specifying the number of columns in the facet.}

\item{byrow}{A logical value indicating whether to fill the plots by row.}
}
\description{
Common arguments for atomic plots
}
\keyword{internal}
