% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotSingle}
\alias{BarPlotSingle}
\alias{BarPlotGrouped}
\alias{BarPlotAtomic}
\alias{BarPlot}
\title{Bar plot}
\usage{
BarPlotSingle(
  data,
  x,
  y,
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  x_text_angle = 0,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  expand = 0,
  fill_by_x = TRUE,
  width = 0.9,
  ...
)

BarPlotGrouped(
  data,
  x,
  y,
  group_by,
  group_by_sep = "_",
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  add_bg = FALSE,
  bg_palette = "Paired",
  bg_palcolor = NULL,
  bg_alpha = 0.2,
  alpha = 1,
  x_text_angle = 0,
  aspect.ratio = 1,
  position = "auto",
  position_dodge_preserve = "total",
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  expand = c(0.05, 0.2, 0, 0.2),
  width = 0.9,
  ...
)

BarPlotAtomic(
  data,
  x,
  y,
  group_by = NULL,
  fill_by_x_if_no_group = TRUE,
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  x_text_angle = 0,
  aspect.ratio = 1,
  position = "auto",
  position_dodge_preserve = "total",
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  expand = NULL,
  width = 0.9,
  facet = TRUE,
  facet_by = NULL,
  facet_scales = "fixed",
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  ...
)

BarPlot(
  data,
  x,
  y,
  group_by = NULL,
  group_by_sep = "_",
  split_by = NULL,
  split_by_sep = "_",
  facet = FALSE,
  facet_scales = "fixed",
  fill_by_x_if_no_group = TRUE,
  add_bg = FALSE,
  bg_palette = "Paired",
  bg_palcolor = NULL,
  bg_alpha = 0.2,
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  x_text_angle = 0,
  aspect.ratio = 1,
  position = "auto",
  position_dodge_preserve = "total",
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  expand = NULL,
  width = 0.9,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  seed = 8525,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A character string specifying the column name of the data frame to plot for the x-axis.}

\item{y}{A character string specifying the column name of the data frame to plot for the y-axis.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{alpha}{A numeric value specifying the transparency of the plot.}

\item{x_text_angle}{A numeric value specifying the angle of the x-axis text.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{keep_empty}{A logical value indicating whether to keep empty groups.
If FALSE, empty groups will be removed.}

\item{expand}{The values to expand the x and y axes. It is like CSS padding.
When a single value is provided, it is used for both axes on both sides.
When two values are provided, the first value is used for the top/bottom side and the second value is used for the left/right side.
When three values are provided, the first value is used for the top side, the second value is used for the left/right side, and the third value is used for the bottom side.
When four values are provided, the values are used for the top, right, bottom, and left sides, respectively.
The values will be applied as 'mult' to the 'expansion' function.
See also \url{https://ggplot2.tidyverse.org/reference/expansion.html}}

\item{fill_by_x}{A logical value indicating whether to fill the bars by the x-axis values.
If FALSE, the bars will be filled a single color (the first color in the palette).}

\item{width}{A numeric value specifying the width of the bars.}

\item{...}{Additional arguments.}

\item{group_by}{Columns to group the data for plotting
For those plotting functions that do not support multiple groups,
They will be concatenated into one column, using \code{\link{group_by_sep}} as the separator}

\item{group_by_sep}{The separator for multiple group_by columns. See \code{\link{group_by}}}

\item{add_bg}{A logical value indicating whether to add a background to the plot.}

\item{bg_palette}{A character string indicating the palette to use for the background.}

\item{bg_palcolor}{A character string indicating the color to use for the background.}

\item{bg_alpha}{A numeric value indicating the alpha of the background.}

\item{position}{A character string indicating the position of the bars.
If "auto", the position will be "stack" if group_by has more than 5 levels, otherwise "dodge".
"fill" is also a valid option. Only works when group_by is not NULL.}

\item{position_dodge_preserve}{Should dodging preserve the "total" width of all elements at a position, or the width of a "single" element?}

\item{fill_by_x_if_no_group}{A logical value indicating whether to fill the bars by the x-axis values if there is no group_by.}

\item{facet}{A logical value indicating whether to facet the plot.}

\item{facet_by}{A column name to facet the plot by.}

\item{facet_scales}{A character string indicating the scales of the facets.}

\item{nrow}{An integer value indicating the number of rows using facet_wrap.}

\item{ncol}{An integer value indicating the number of columns using facet_wrap.}

\item{byrow}{A logical value indicating whether to fill the facets by row.}

\item{split_by}{The column(s) to split data by and plot separately or facet by.
When \code{facet} is TRUE, the columns specified here will be used to facet the plot
When \code{facet} is FALSE, the columns specified here will be used to split the data
and generate multiple plots and combine them into one
If multiple columns are specified
When \code{facet} is TRUE, up to 2 columns are allowed. If one column,
\code{ggplot2::facet_wrap} is used, the number of rows and columns is determined
by \code{nrow} and \code{ncol}
If two columns, \code{ggplot2::facet_grid} is used. The first column is used for rows
and the second column is used for columns
When \code{facet} is FALSE, a warning will be issued and the columns will be
concatenated into one column, using \code{split_by_sep} as the separator}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{combine}{Whether to combine the plots into one when facet is FALSE. Default is TRUE.}

\item{seed}{The random seed to use. Default is 8525.}
}
\value{
A ggplot object.

A ggplot object.

A ggplot object.

A ggplot object
}
\description{
Provides a base function to create bar plots.
}
\keyword{internal}
