% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldim_plot.R
\name{CellDimPlotAtomic}
\alias{CellDimPlotAtomic}
\title{CellDimPlotAtomic}
\usage{
CellDimPlotAtomic(
  x,
  group_by = NULL,
  group_by_sep = "_",
  split_by = NULL,
  split_by_sep = "_",
  seed = 8525,
  facet = FALSE,
  guess_size = FALSE,
  res = 100,
  facet_scales = "fixed",
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  reduction = NULL,
  bg_color = "grey80",
  pt_size = NULL,
  pt_alpha = 1,
  add_mark = FALSE,
  mark_type = c("ellipse", "hull", "rect", "circle"),
  cells_highlight = NULL,
  cols_highlight = "black",
  sizes_highlight = 1,
  alpha_highlight = 1,
  stroke_highlight = 0.5,
  dims = c(1, 2),
  show_na = FALSE,
  show_stat = !identical(theme, "theme_blank") && !facet,
  mark_expand = unit(3, "mm"),
  mark_alpha = 0.1,
  mark_linetype = 1,
  stat_pie_label = FALSE,
  label = FALSE,
  label_insitu = FALSE,
  label_size = 4,
  label_fg = "white",
  label_bg = "black",
  label_bg_r = 0.1,
  label_repel = FALSE,
  label_repulsion = 20,
  label_point_size = 1,
  label_point_color = "black",
  label_segment_color = "black",
  add_density = FALSE,
  density_color = "grey80",
  density_filled = FALSE,
  density_filled_palette = "Greys",
  density_filled_palcolor = NULL,
  lineages = NULL,
  lineages_trim = c(0.01, 0.99),
  lineages_span = 0.75,
  lineages_palette = "Dark2",
  lineages_palcolor = NULL,
  lineages_arrow = arrow(length = unit(0.1, "inches")),
  lineages_linewidth = 1,
  lineages_line_bg = "white",
  lineages_line_bg_stroke = 0.5,
  lineages_whiskers = FALSE,
  lineages_whiskers_linewidth = 0.5,
  lineages_whiskers_alpha = 0.5,
  stat_by = NULL,
  stat_frac = "none",
  stat_plot_type = "pie",
  stat_plot_size = 0.15,
  stat_plot_palette = "Set1",
  stat_palcolor = NULL,
  stat_plot_position = c("stack", "dodge"),
  stat_plot_alpha = 1,
  stat_plot_label = FALSE,
  stat_plot_label_size = 3,
  graph = NULL,
  edge_size = c(0.05, 0.5),
  edge_alpha = 0.1,
  edge_color = "grey40",
  paga = NULL,
  paga_type = "connectivities",
  paga_node_size = 4,
  paga_edge_threshold = 0.01,
  paga_edge_size = c(0.2, 1),
  paga_edge_color = "grey40",
  paga_edge_alpha = 0.5,
  paga_show_transition = FALSE,
  paga_transition_threshold = 0.01,
  paga_transition_size = c(0.2, 1),
  paga_transition_color = "black",
  paga_transition_alpha = 1,
  velocity = NULL,
  velocity_plot_type = "raw",
  velocity_n_neighbors = ceiling(ncol(x@assays[[1]])/50),
  velocity_density = 1,
  velocity_smooth = 0.5,
  velocity_scale = 1,
  velocity_min_mass = 1,
  velocity_cutoff_perc = 5,
  velocity_arrow_color = "black",
  velocity_arrow_angle = 20,
  streamline_L = 5,
  streamline_minL = 1,
  streamline_res = 1,
  streamline_n = 15,
  streamline_width = c(0, 0.8),
  streamline_alpha = 1,
  streamline_color = NULL,
  streamline_palette = "RdYlBu",
  streamline_palcolor = NULL,
  streamline_bg_color = "white",
  streamline_bg_stroke = 0.5,
  hex = FALSE,
  hex_count = TRUE,
  hex_bins = 50,
  hex_binwidth = NULL,
  hex_linewidth = 0.5,
  raster = NULL,
  raster_dpi = c(512, 512),
  ...
)
}
\arguments{
\item{x}{Seurat object}

\item{group_by}{Name of one or more meta.data columns to group (color) cells by (for example, seurat_clusters).}

\item{group_by_sep}{The separator for multiple group_by columns. See \code{\link{group_by}}}

\item{split_by}{The column(s) to split data by and plot separately or facet by.
When \code{facet} is TRUE, the columns specified here will be used to facet the plot
When \code{facet} is FALSE, the columns specified here will be used to split the data
and generate multiple plots and combine them into one
If multiple columns are specified
When \code{facet} is TRUE, up to 2 columns are allowed. If one column,
\code{ggplot2::facet_wrap} is used, the number of rows and columns is determined
by \code{nrow} and \code{ncol}
If two columns, \code{ggplot2::facet_grid} is used. The first column is used for rows
and the second column is used for columns
When \code{facet} is FALSE, a warning will be issued and the columns will be
concatenated into one column, using \code{split_by_sep} as the separator}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{seed}{The random seed to use. Default is 8525.}

\item{facet}{Whether to facet the plot. Default is FALSE.
Otherwise, the data will be split by \code{\link{split_by}} and generate multiple plots
and combine them into one using \code{patchwork::wrap_plots}}

\item{facet_scales}{Whether to scale the axes of facets. Default is "fixed"
Other options are "free", "free_x", "free_y". See \code{\link{ggplot2::facet_wrap}}}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{nrow}{The number of rows in facet_wrap/wrap_plots. Default is NULL.}

\item{ncol}{The number of columns in facet_wrap/wrap_plots. Default is NULL.}

\item{byrow}{Whether to fill the plots by row. Default is TRUE.}

\item{reduction}{Name of the reduction to plot (for example, "umap").}

\item{bg_color}{Color value for background(NA) points.}

\item{pt_size}{Point size.}

\item{pt_alpha}{Point transparency.}

\item{add_mark}{Whether to add a mark layer on the plot.}

\item{mark_type}{Type of mark to add to the plot. Default is "hull".}

\item{cells_highlight}{A vector of cell names to highlight or an expression passed to \code{\link{subset}} to select cells to highlight.}

\item{cols_highlight}{Color used to highlight the cells.}

\item{sizes_highlight}{Size of highlighted cell points.}

\item{alpha_highlight}{Transparency of highlighted cell points.}

\item{stroke_highlight}{Border width of highlighted cell points.}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{show_na}{Whether to assign a color from the color palette to NA group. If \code{FALSE}, cell points with NA level will colored by \code{bg_color}.}

\item{show_stat}{Whether to show statistical information on the plot.}

\item{mark_expand}{Expansion of the mark.}

\item{mark_alpha}{Transparency of the mark.}

\item{mark_linetype}{Line type of the mark.}

\item{stat_pie_label}{Whether to add labels in the pie plot.}

\item{label}{Whether to label the cell groups.}

\item{label_insitu}{Whether to place the raw labels (group names) in the center of the cells with the corresponding group. Default is \code{FALSE}, which using numbers instead of raw labels.}

\item{label_size}{Size of labels.}

\item{label_fg}{Foreground color of label.}

\item{label_bg}{Background color of label.}

\item{label_bg_r}{Background ratio of label.}

\item{label_repel}{Logical value indicating whether the label is repel away from the center points.}

\item{label_repulsion}{Force of repulsion between overlapping text labels. Defaults to 20.}

\item{label_point_size}{Size of the center points.}

\item{label_point_color}{Color of the center points.}

\item{label_segment_color}{Color of the line segment for labels.}

\item{add_density}{Whether to add a density layer on the plot.}

\item{density_color}{Color of the density contours lines.}

\item{density_filled}{Whether to add filled contour bands instead of contour lines.}

\item{density_filled_palette}{Color palette used to fill contour bands.}

\item{density_filled_palcolor}{Custom colors used to fill contour bands.}

\item{lineages}{Lineages/pseudotime to add to the plot. If specified, curves will be fitted using \code{\link[stats]{loess}} method.}

\item{lineages_trim}{Trim the leading and the trailing data in the lineages.}

\item{lineages_span}{The parameter α which controls the degree of smoothing in \code{\link[stats]{loess}} method.}

\item{lineages_palette}{Color palette used for lineages.}

\item{lineages_palcolor}{Custom colors used for lineages.}

\item{lineages_arrow}{Set arrows of the lineages. See \code{\link[grid]{arrow}}.}

\item{lineages_linewidth}{Width of fitted curve lines for lineages.}

\item{lineages_line_bg}{Background color of curve lines for lineages.}

\item{lineages_line_bg_stroke}{Border width of curve lines background.}

\item{lineages_whiskers}{Whether to add whiskers for lineages.}

\item{lineages_whiskers_linewidth}{Width of whiskers for lineages.}

\item{lineages_whiskers_alpha}{Transparency of whiskers for lineages.}

\item{stat_by}{The name of a metadata column to stat.}

\item{stat_frac}{The way of calculating the fraction. Default is "none".
Possible values are "group", "ident", "cluster", "all", "none".
\itemize{
\item group: calculate the fraction in each group.
The total fraction of the cells of idents in each group will be 1.
When \code{group-by} is not specified, it will be the same as \code{all}.
\item ident: calculate the fraction in each ident.
The total fraction of the cells of groups in each ident will be 1.
Only works when \code{group-by} is specified.
\item cluster: alias of \code{ident}.
\item all: calculate the fraction against all cells.
\item none: do not calculate the fraction, use the number of cells instead.
}}

\item{stat_plot_type}{Set the statistical plot type.}

\item{stat_plot_size}{Set the statistical plot size. Defaults to 0.1}

\item{stat_plot_palette}{Color palette used in statistical plot.}

\item{stat_palcolor}{Custom colors used in statistical plot}

\item{stat_plot_position}{Position adjustment in statistical plot.}

\item{stat_plot_alpha}{Transparency of the statistical plot.}

\item{stat_plot_label}{Whether to add labels in the statistical plot.}

\item{stat_plot_label_size}{Label size in the statistical plot.}

\item{graph}{Specify the graph name to add edges between cell neighbors to the plot.}

\item{edge_size}{Size of edges.}

\item{edge_alpha}{Transparency of edges.}

\item{edge_color}{Color of edges.}

\item{paga}{Specify the calculated paga results to add a PAGA graph layer to the plot.}

\item{paga_type}{PAGA plot type. "connectivities" or "connectivities_tree".}

\item{paga_node_size}{Size of the nodes in PAGA plot.}

\item{paga_edge_threshold}{Threshold of edge connectivities in PAGA plot.}

\item{paga_edge_size}{Size of edges in PAGA plot.}

\item{paga_edge_color}{Color of edges in PAGA plot.}

\item{paga_edge_alpha}{Transparency of edges in PAGA plot.}

\item{paga_show_transition}{Whether to show transitions between edges.}

\item{paga_transition_threshold}{Threshold of transition edges in PAGA plot.}

\item{paga_transition_size}{Size of transition edges in PAGA plot.}

\item{paga_transition_color}{Color of transition edges in PAGA plot.}

\item{paga_transition_alpha}{Transparency of transition edges in PAGA plot.}

\item{velocity}{Specify the calculated RNA velocity mode to add a velocity layer to the plot.}

\item{velocity_plot_type}{Set the velocity plot type.}

\item{velocity_n_neighbors}{Set the number of neighbors used in velocity plot.}

\item{velocity_density}{Set the density value used in velocity plot.}

\item{velocity_smooth}{Set the smooth value used in velocity plot.}

\item{velocity_scale}{Set the scale value used in velocity plot.}

\item{velocity_min_mass}{Set the min_mass value used in velocity plot.}

\item{velocity_cutoff_perc}{Set the cutoff_perc value used in velocity plot.}

\item{velocity_arrow_color}{Color of arrows in velocity plot.}

\item{velocity_arrow_angle}{Angle of arrows in velocity plot.}

\item{streamline_L}{Typical length of a streamline in x and y units}

\item{streamline_minL}{Minimum length of segments to show.}

\item{streamline_res}{Resolution parameter (higher numbers increases the resolution).}

\item{streamline_n}{Number of points to draw.}

\item{streamline_width}{Size of streamline.}

\item{streamline_alpha}{Transparency of streamline.}

\item{streamline_color}{Color of streamline.}

\item{streamline_palette}{Color palette used for streamline.}

\item{streamline_palcolor}{Custom colors used for streamline.}

\item{streamline_bg_color}{Background color of streamline.}

\item{streamline_bg_stroke}{Border width of streamline background.}

\item{hex}{Whether to chane the plot type from point to the hexagonal bin.}

\item{hex_count}{Whether show cell counts in each hexagonal bin.}

\item{hex_bins}{Number of hexagonal bins.}

\item{hex_binwidth}{Hexagonal bin width.}

\item{hex_linewidth}{Border width of hexagonal bins.}

\item{raster}{Convert points to raster format, default is NULL which automatically rasterizes if plotting more than 100,000 cells}

\item{raster_dpi}{Pixel resolution for rasterized plots, passed to geom_scattermore(). Default is c(512, 512).}

\item{...}{Additional arguments}
}
\description{
CellDimPlotAtomic
}
\keyword{internal}
