% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cellphonedb_res}
\alias{cellphonedb_res}
\title{A toy example of CellPhoneDB output from LIANA}
\format{
A \code{data.frame} with 10 rows and 6 columns
}
\source{
\url{https://liana-py.readthedocs.io/en/latest/notebooks/basic_usage.html#Tileplot}
}
\description{
The dataset is generated using python package LIANA
}
\examples{
\dontrun{
# Python code
# # import liana
# import liana as li
# # needed for visualization and toy data
# import scanpy as sc
#
# from liana.method import cellphonedb
# adata = sc.datasets.pbmc68k_reduced()
# cellphonedb(adata,
#             groupby='bulk_labels',
#             # NOTE by default the resource uses HUMAN gene symbols
#             resource_name='consensus',
#             expr_prop=0.1,
#             verbose=True, key_added='cpdb_res')
# cellphonedb_res = adata.uns['cpdb_res']
# cellphonedb_res = cellphonedb_res[cellphonedb_res['cellphone_pvals'] < 0.05]
}
}
\concept{data}
