% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellstat_plot.R
\name{CellStatPlotAtomic.bar}
\alias{CellStatPlotAtomic.bar}
\title{CellStatPlotAtomic method for bar plot}
\usage{
\method{CellStatPlotAtomic}{bar}(
  x,
  ident = "seurat_clusters",
  frac = "none",
  flip = FALSE,
  bar_position = "auto",
  pie_label = TRUE,
  group_by = NULL,
  group_by_sep = "_",
  split_by = NULL,
  split_by_sep = "_",
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  keep_empty = FALSE,
  x_text_angle = 90,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  theme = "theme_scp",
  theme_args = list(),
  facet = FALSE,
  facet_scales = "fixed",
  res = 100,
  guess_size = FALSE,
  seed = 8525,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  circos_labels_rot = FALSE,
  ...
)
}
\arguments{
\item{x}{The data frame or Seurat object}

\item{ident}{The column with the cell identities. i.e. clusters. Default: seurat_clusters}

\item{frac}{The way of calculating the fraction. Default is "none".
Possible values are "group", "ident", "cluster", "all", "none".
\itemize{
\item group: calculate the fraction in each group.
The total fraction of the cells of idents in each group will be 1.
When \code{group-by} is not specified, it will be the same as \code{all}.
\item ident: calculate the fraction in each ident.
The total fraction of the cells of groups in each ident will be 1.
Only works when \code{group-by} is specified.
\item cluster: alias of \code{ident}.
\item all: calculate the fraction against all cells.
\item none: do not calculate the fraction, use the number of cells instead.
}}

\item{flip}{Whether to flip the cluster and group, that is,
using group as the x-axis and cluster to fill the plot.
For circos plot, when transposed, the arrows will be drawn from the idents (by \code{ident}) to the
the groups (by \code{group_by}).
Only works when \code{group_by} is specified.}

\item{bar_position}{The position of the bars. Default is "auto".
\itemize{
\item stack: Use \code{position_stack()}.
\item fill: Use \code{position_fill()}.
\item dodge: Use \code{position_dodge()}.
\item auto: Use \code{stack} when there are more than 5 groups, otherwise use \code{dodge}.
}}

\item{pie_label}{Whether to add labels to the pie plot. Default is TRUE.}

\item{group_by}{Columns to group the data for plotting
For those plotting functions that do not support multiple groups,
They will be concatenated into one column, using \code{\link{group_by_sep}} as the separator}

\item{group_by_sep}{The separator for multiple group_by columns. See \code{\link{group_by}}}

\item{split_by}{The column(s) to split data by and plot separately or facet by.
When \code{facet} is TRUE, the columns specified here will be used to facet the plot
When \code{facet} is FALSE, the columns specified here will be used to split the data
and generate multiple plots and combine them into one
If multiple columns are specified
When \code{facet} is TRUE, up to 2 columns are allowed. If one column,
\code{ggplot2::facet_wrap} is used, the number of rows and columns is determined
by \code{nrow} and \code{ncol}
If two columns, \code{ggplot2::facet_grid} is used. The first column is used for rows
and the second column is used for columns
When \code{facet} is FALSE, a warning will be issued and the columns will be
concatenated into one column, using \code{split_by_sep} as the separator}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{alpha}{A numeric value specifying the transparency of the plot.}

\item{keep_empty}{A logical value indicating whether to keep empty groups.
If FALSE, empty groups will be removed.}

\item{x_text_angle}{A numeric value specifying the angle of the x-axis text.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{facet}{Whether to facet the plot. Default is FALSE.
Otherwise, the data will be split by \code{\link{split_by}} and generate multiple plots
and combine them into one using \code{patchwork::wrap_plots}}

\item{facet_scales}{Whether to scale the axes of facets. Default is "fixed"
Other options are "free", "free_x", "free_y". See \code{\link{ggplot2::facet_wrap}}}

\item{seed}{The random seed to use. Default is 8525.}

\item{nrow}{The number of rows in facet_wrap/wrap_plots. Default is NULL.}

\item{ncol}{The number of columns in facet_wrap/wrap_plots. Default is NULL.}

\item{byrow}{Whether to fill the plots by row. Default is TRUE.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{circos_labels_rot}{Whether to rotate the labels in the circos plot.
In case the labels are too long. Default is FALSE.}

\item{...}{Additional arguments}
}
\description{
CellStatPlotAtomic method for bar plot
}
\keyword{internal}
