% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellstat_plot.R
\name{CellStatPlotAtomic.default}
\alias{CellStatPlotAtomic.default}
\title{Default method for \code{CellStatPlotAtomic}}
\usage{
\method{CellStatPlotAtomic}{default}(x, ...)
}
\description{
Default method for \code{CellStatPlotAtomic}
}
\keyword{internal}
