% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellstat_plot.R
\name{CellStatPlotAtomic}
\alias{CellStatPlotAtomic}
\title{CellStatPlotAtomic}
\usage{
CellStatPlotAtomic(x, ...)
}
\arguments{
\item{x}{The data frame or Seurat object}

\item{...}{Additional arguments}
}
\value{
A ggplot object or a list with the plot and the height and width of the plot if guess_size is TRUE
}
\description{
CellStatPlotAtomic
}
\keyword{internal}
