% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonaldivplot.R
\name{ClonalDiversity}
\alias{ClonalDiversity}
\title{Calculate the clonal diversities.}
\usage{
ClonalDiversity(
  input.data,
  cloneCall = "gene",
  chain = "both",
  method = c("shannon", "gini.coeff", "inv.simpson", "norm.entropy", "gini.simpson",
    "chao1", "ACE"),
  group_by = NULL,
  n_boots = 100
)
}
\description{
Calculate the clonal diversities.
}
\keyword{internal}
