% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalstatplot.R
\name{ClonalDynamicsPlot}
\alias{ClonalDynamicsPlot}
\title{ClonalDynamicsPlot}
\usage{
ClonalDynamicsPlot(data, ...)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or
\link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{...}{Other arguments.}
}
\description{
This function is deprecated. Please use \code{\link[=ClonalStatPlot]{ClonalStatPlot()}} instead.
}
