% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_dimreduc<-}
\alias{default_dimreduc<-}
\title{Set default dimension reduction for Seurat object}
\usage{
default_dimreduc(object) <- value
}
\arguments{
\item{object}{Seurat object to modify}

\item{value}{Name of the default dimension reduction to set}
}
\value{
Modified Seurat object with updated default dimension reduction
}
\description{
Set default dimension reduction for Seurat object
}
\details{
\verb{DefaultDimReduc<-} was just introduced in Seurat v5.4.0
See: https://github.com/satijalab/seurat-object/pull/268
and https://github.com/satijalab/seurat-object/releases/tag/v5.4.0
This function is a fallback for older Seurat versions that do not have this setter function.
It will simply save the default dimension reduction name in the Seurat object's misc slot under \code{DefaultDimReduc}.
When the setter function is available, it will use that instead.
}
\keyword{internal}
