% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_dimreduc}
\alias{default_dimreduc}
\title{Get default dimension reduction for Seurat object}
\usage{
default_dimreduc(object)
}
\arguments{
\item{object}{Seurat object to query}
}
\value{
Name of the default dimension reduction, or NULL if not set
}
\description{
Get default dimension reduction for Seurat object
}
\details{
\code{DefaultDimReduc} was just introduced in Seurat v5.4.0.
This function will first check if we have \code{DefaultDimReduc} value in the misc slot (set by our fallback setter),
and if not, it will try to call the official \code{DefaultDimReduc} getter function. If that also fails, it will return NULL.
}
\keyword{internal}
