% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_call}
\alias{do_call}
\title{Call a function with a list of arguments}
\usage{
do_call(what, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{what}{either a function or a non-empty character string naming the
    function to be called.}

\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{quote}{a logical value indicating whether to quote the
    arguments.}

\item{envir}{an environment within which to evaluate the call.  This
    will be most useful if \code{what} is a character string and
    the arguments are symbols or quoted expressions.}
}
\value{
The result of the function call
}
\description{
Different from \code{do.call}, this is a faster version, especially when there are
big objects in the list.
}
\keyword{internal}
