% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{.df_to_polygon_spatvector}
\alias{.df_to_polygon_spatvector}
\title{Convert a data.frame to a terra SpatVector polygon}
\usage{
.df_to_polygon_spatvector(df, id, x, y, attrs = NULL)
}
\description{
Convert a data.frame to a terra SpatVector polygon
}
\keyword{internal}
