% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialutils.R
\name{.new_scale_layers}
\alias{.new_scale_layers}
\title{Add new scale layers}
\usage{
.new_scale_layers(scales)
}
\arguments{
\item{scales}{The scales have been already applied in the previous layers
(so that we need to add new scales only for the new layers).}
}
\value{
A list of ggnewscale layers for the specified scales.
}
\description{
Add new scale layers
}
\keyword{internal}
