% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialutils.R
\name{.points_args}
\alias{.points_args}
\title{Get points arguments from the dot-dot-dot arguments in the spatial plot function}
\usage{
.points_args(args, ...)
}
\arguments{
\item{args}{A list of arguments (the dot arguments) passed to the spatial plot function.}

\item{...}{Additional arguments that will be parsed.}
}
\value{
A list of points arguments.
}
\description{
Get points arguments from the dot-dot-dot arguments in the spatial plot function
}
\keyword{internal}
