% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialutils.R
\name{.rect_bg_image}
\alias{.rect_bg_image}
\title{Draw a rectangle as the image (background) of a spatial plot and fill with a given color}
\usage{
.rect_bg_image(args, fill)
}
\arguments{
\item{args}{A list of arguments (the dot arguments) passed to the spatial plot function.}

\item{fill}{The color to fill the rectangle.}
}
\value{
A ggplot2 layer object that can be added to a ggplot.
}
\description{
Draw a rectangle as the image (background) of a spatial plot and fill with a given color
}
\keyword{internal}
