% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{.seurat_points_layer}
\alias{.seurat_points_layer}
\title{Process points layer for Seurat spatial plots}
\usage{
.seurat_points_layer(
  object,
  points_x = "x",
  points_y = "y",
  image,
  args,
  crop,
  points_data,
  ext_unscaled,
  scale_factor,
  group_by,
  features,
  layer,
  legend.position,
  legend.direction,
  label,
  label_size,
  label_fg,
  label_bg,
  label_bg_r,
  label_repel,
  label_repulsion,
  label_pt_size,
  label_pt_color,
  label_segment_color,
  label_insitu,
  highlight,
  highlight_alpha,
  highlight_size,
  highlight_color,
  highlight_stroke,
  palette,
  palette_reverse,
  palcolor,
  flip_y,
  ext
)
}
\description{
Process points layer for Seurat spatial plots
}
\keyword{internal}
