% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalstatplot.R
\name{DummyClonalScatterPlot}
\alias{DummyClonalScatterPlot}
\title{DummyClonalScatterPlot}
\usage{
DummyClonalScatterPlot(df, title, group_by, scatter_cor, size_by, ...)
}
\arguments{
\item{df}{The data frame with the clonal data.
The data frame should have the columns: group_by, 'count', 'fraction'.}
}
\value{
A ggplot object
}
\description{
Function to plot the scatter plot of the clonal data for a dummy group pair.
}
\keyword{internal}
