% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{EnrichmentPlotAtomic.default}
\alias{EnrichmentPlotAtomic.default}
\title{EnrichmentPlotAtomic for default}
\usage{
\method{EnrichmentPlotAtomic}{default}(x, ...)
}
\description{
EnrichmentPlotAtomic for default
}
\keyword{internal}
