% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{EnrichmentPlotAtomic.network}
\alias{EnrichmentPlotAtomic.network}
\title{EnrichmentPlotAtomic for "network" plot}
\usage{
\method{EnrichmentPlotAtomic}{network}(
  x,
  enrichtool = NULL,
  db = NULL,
  metric = "p.adjust",
  metric_value = 0.05,
  group_by = NULL,
  group_by_sep = "_",
  split_by = "Database",
  split_by_sep = "_",
  pvalue_cutoff = NULL,
  padjust_cutoff = 0.05,
  bar_showcutoff = TRUE,
  top_term = 10,
  compare_only_sig = FALSE,
  top_word = 100,
  word_type = c("term", "feature"),
  word_size = c(2, 8),
  words_excluded = scplotter::words_excluded,
  network_layout = "fr",
  network_labelsize = 5,
  network_blendmode = "blend",
  network_layoutadjust = TRUE,
  network_adjscale = 60,
  network_adjiter = 100,
  enrichmap_layout = "fr",
  enrichmap_cluster = "fast_greedy",
  enrichmap_label = "term",
  enrichmap_labelsize = 5,
  enrichmap_nlabel = 4,
  enrichmap_show_keyword = FALSE,
  enrichmap_mark = "ellipse",
  enrichmap_expand = c(0.5, 0.5),
  x_text_angle = 0,
  character_width = 50,
  lineheight = 0.8,
  palette = "Spectral",
  palcolor = NULL,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  theme = "theme_scp",
  theme_args = list(),
  facet = FALSE,
  facet_scales = "fixed",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  seed = 8525,
  res = 100,
  guess_size = FALSE,
  ...
)
}
\arguments{
\item{x}{A data frame with Enrichment results}

\item{enrichtool}{The enrichment tool used to generate the enrichment results. Default is NULL.
Will automatically detect the enrichment tool from the enrichment results.
Currently supported tools are "enrichr" and "clusterprofiler".}

\item{db}{The database to use for enrichment plot. Default: all Database in the enrichment results.
If Database is not specified in the enrichment results, the default value is this value.}

\item{metric}{The metric to use for coloring the nodes. Default is "p.adjust".}

\item{metric_value}{The metric value to use for coloring the nodes. Default is 0.05.}

\item{group_by}{The grouping variable(s) for the plot. Default is NULL.
Only work when \code{plot_type} is "bar" or "comparison".
For bar plot, bars will be colored by this variable.}

\item{group_by_sep}{The separator for multiple group_by columns. See \code{\link{group_by}}}

\item{split_by}{The column(s) to split data by and plot separately or facet by.
When \code{facet} is TRUE, the columns specified here will be used to facet the plot
When \code{facet} is FALSE, the columns specified here will be used to split the data
and generate multiple plots and combine them into one
If multiple columns are specified
When \code{facet} is TRUE, up to 2 columns are allowed. If one column,
\code{ggplot2::facet_wrap} is used, the number of rows and columns is determined
by \code{nrow} and \code{ncol}
If two columns, \code{ggplot2::facet_grid} is used. The first column is used for rows
and the second column is used for columns
When \code{facet} is FALSE, a warning will be issued and the columns will be
concatenated into one column, using \code{split_by_sep} as the separator}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{pvalue_cutoff}{The p-value cutoff. Default is NULL. Only work when \code{padjust_cutoff} is NULL.}

\item{padjust_cutoff}{The p-adjusted cutoff. Default is 0.05.}

\item{bar_showcutoff}{Whether to show the p-value or p-adjusted cutoff line. Default is TRUE.}

\item{top_term}{The number of top terms to display. Default is 6, or 100 if 'plot_type' is "enrichmap".}

\item{compare_only_sig}{Whether to compare only significant terms. Default is FALSE.}

\item{top_word}{The number of top words to display for wordcloud. Default is 100.}

\item{word_type}{The type of words to display in wordcloud. Options are "term" and "feature". Default is "term".}

\item{word_size}{The size range for words in wordcloud. Default is c(2, 8).}

\item{words_excluded}{Words to be excluded from the wordcloud. The default value is NULL, which means that the built-in words (scplotter::words_excluded) will be used.}

\item{network_layout}{The layout algorithm to use for network plot. Options are "fr", "kk","random", "circle", "tree", "grid", or other algorithm from 'igraph' package. Default is "fr".}

\item{network_labelsize}{The label size for network plot. Default is 5.}

\item{network_blendmode}{The blend mode for network plot. Default is "blend".}

\item{network_layoutadjust}{Whether to adjust the layout of the network plot to avoid overlapping words. Default is TRUE.}

\item{network_adjscale}{The scale for adjusting network plot layout. Default is 60.}

\item{network_adjiter}{The number of iterations for adjusting network plot layout. Default is 100.}

\item{enrichmap_layout}{The layout algorithm to use for enrichmap plot. Options are "fr", "kk","random", "circle", "tree", "grid", or other algorithm from 'igraph' package. Default is "fr".}

\item{enrichmap_cluster}{The clustering algorithm to use for enrichmap plot. Options are "walktrap", "fast_greedy", or other algorithm from 'igraph' package. Default is "fast_greedy".}

\item{enrichmap_label}{The label type for enrichmap plot. Options are "term" and "feature". Default is "term".}

\item{enrichmap_labelsize}{The label size for enrichmap plot. Default is 5.}

\item{enrichmap_show_keyword}{Whether to show the keyword of terms or features in enrichmap plot. Default is FALSE.}

\item{enrichmap_mark}{The mark shape for enrichmap plot. Options are "ellipse" and "hull". Default is "ellipse".}

\item{enrichmap_expand}{The expansion factor for enrichmap plot. Default is c(0.5, 0.5).}

\item{x_text_angle}{A numeric value specifying the angle of the x-axis text.}

\item{character_width}{The maximum width of character of descriptions. Default is 50.}

\item{lineheight}{The line height for y-axis labels. Default is 0.5.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{facet}{Whether to facet the plot. Default is FALSE.
Otherwise, the data will be split by \code{\link{split_by}} and generate multiple plots
and combine them into one using \code{patchwork::wrap_plots}}

\item{facet_scales}{Whether to scale the axes of facets. Default is "fixed"
Other options are "free", "free_x", "free_y". See \code{\link{ggplot2::facet_wrap}}}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{nrow}{The number of rows in facet_wrap/wrap_plots. Default is NULL.}

\item{ncol}{The number of columns in facet_wrap/wrap_plots. Default is NULL.}

\item{byrow}{Whether to fill the plots by row. Default is TRUE.}

\item{seed}{The random seed to use. Default is 8525.}

\item{...}{Additional arguments for future expansion.}
}
\description{
EnrichmentPlotAtomic for "network" plot
}
\keyword{internal}
