% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{enrichr_multidb_example}
\alias{enrichr_multidb_example}
\title{An example results from enrichr with multiple databases}
\description{
The variable "enrichr_multidb_example" represents an example results from enrichr with multiple databases.
}
\concept{data}
