% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_expansion}
\alias{expand_expansion}
\title{Expand the plot area with CSS-like padding}
\usage{
expand_expansion(expand)
}
\arguments{
\item{expand}{A numeric vector of length 1, 2, 3, or 4}
}
\value{
A list with x and y values for expand
}
\description{
Expand the plot area with CSS-like padding
}
\keyword{internal}
