% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{guess_enrichment_tool}
\alias{guess_enrichment_tool}
\title{Guess_enrichment_tool}
\usage{
guess_enrichment_tool(x)
}
\arguments{
\item{x}{A data frame with Enrichment results}
}
\value{
The enrichment tool used to generate the enrichment results.
}
\description{
Guess the enrichment tool from the enrichment results.
}
\keyword{internal}
