% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h5group_to_dataframe}
\alias{h5group_to_dataframe}
\title{Try to read a H5Group as a data.frame}
\usage{
h5group_to_dataframe(x)
}
\arguments{
\item{x}{H5Group object}
}
\value{
A data.frame if successful
}
\description{
Try to read a H5Group as a data.frame
}
\keyword{internal}
