% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h5group_to_matrix}
\alias{h5group_to_matrix}
\title{Try to read a H5Group as a matrix}
\usage{
h5group_to_matrix(x)
}
\arguments{
\item{x}{H5Group object}
}
\value{
A sparseMatrix if successful
}
\description{
Try to read a H5Group as a matrix
}
\keyword{internal}
