% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_plot.R
\name{LineagePlot}
\alias{LineagePlot}
\title{LineagePlot}
\usage{
LineagePlot(
  x,
  lineages,
  reduction = NULL,
  dims = c(1, 2),
  trim = c(0.01, 0.99),
  span = 0.75,
  palette = "Dark2",
  palcolor = NULL,
  lineages_arrow = arrow(length = unit(0.1, "inches")),
  linewidth = 1,
  line_bg = "white",
  line_bg_stroke = 0.5,
  whiskers = FALSE,
  whiskers_linewidth = 0.5,
  whiskers_alpha = 0.5,
  aspect.ratio = 1,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  theme = "theme_scp",
  theme_args = list(),
  return_layer = FALSE,
  seed = 8525,
  guess_size = FALSE,
  res = 100,
  ...
)
}
\arguments{
\item{x}{An object of class Seurat.}

\item{lineages}{A character vector that specifies the lineages to be included. Typically, use the pseudotime of cells.}

\item{reduction}{An optional string specifying the dimensionality reduction method to use.}

\item{dims}{A numeric vector of length 2 specifying the dimensions to plot.}

\item{trim}{A numeric vector of length 2 specifying the quantile range of lineages to include in the plot.}

\item{span}{A numeric value specifying the span of the loess smoother.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{lineages_arrow}{An arrow object specifying the arrow for lineages.}

\item{linewidth}{A numeric value specifying the linewidth for the lineages.}

\item{line_bg}{A character string specifying the color for the background lines.}

\item{line_bg_stroke}{A numeric value specifying the stroke width for the background lines.}

\item{whiskers}{A logical value indicating whether to include whiskers in the plot.}

\item{whiskers_linewidth}{A numeric value specifying the linewidth for the whiskers.}

\item{whiskers_alpha}{A numeric value specifying the transparency for the whiskers.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{return_layer}{A logical value indicating whether to return the plot as a layer.}

\item{seed}{The random seed to use. Default is 8525.}

\item{...}{Additional arguments for future expansion.}
}
\description{
Generate a lineage plot based on the pseudotime.
}
\examples{
data(pancreas_sub)
LineagePlot(pancreas_sub, lineages = paste0("Lineage", 1:3))
LineagePlot(pancreas_sub, lineages = paste0("Lineage", 1:3), whiskers = TRUE)
}
\seealso{
\code{\link{CellDimPlot}}
}
