% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{normalize_enrichment.default}
\alias{normalize_enrichment.default}
\title{Default method for \code{normalize_enrichment}}
\usage{
\method{normalize_enrichment}{default}(x, db, ...)
}
\description{
Default method for \code{normalize_enrichment}
}
\keyword{internal}
