% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{normalize_enrichment.enrichment_clusterprofiler}
\alias{normalize_enrichment.enrichment_clusterprofiler}
\title{Normalize_enrichment for "enrichment_clusterprofiler"}
\usage{
\method{normalize_enrichment}{enrichment_clusterprofiler}(x, db, ...)
}
\description{
Normalize_enrichment for "enrichment_clusterprofiler"
}
\keyword{internal}
