% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{normalize_enrichment.enrichment_enrichr}
\alias{normalize_enrichment.enrichment_enrichr}
\title{Normalize_enrichment for "enrichment_enrichr"}
\usage{
\method{normalize_enrichment}{enrichment_enrichr}(x, db, ...)
}
\description{
Normalize_enrichment for "enrichment_enrichr"
}
\keyword{internal}
