% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{normalize_enrichment}
\alias{normalize_enrichment}
\title{Normalize_enrichment}
\usage{
normalize_enrichment(x, db, ...)
}
\arguments{
\item{x}{A data frame with Enrichment results}
}
\value{
The normalized enrichment results.
}
\description{
Normalize the enrichment results.
}
\keyword{internal}
