% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart.R
\name{PieChartAtomic}
\alias{PieChartAtomic}
\alias{PieChart}
\title{Pie chart}
\usage{
PieChartAtomic(
  data,
  x,
  y,
  label = "y",
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  facet = FALSE,
  facet_by = NULL,
  facet_scales = "fixed",
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  ...
)

PieChart(
  data,
  x,
  y,
  label = "y",
  split_by = NULL,
  split_by_sep = "_",
  facet = FALSE,
  facet_scales = "fixed",
  theme = "theme_scp",
  theme_args = list(),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{label}{Which column to use as the label. NULL means no label.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{alpha}{A numeric value specifying the transparency of the plot.}

\item{facet}{A logical value indicating whether to facet the plot.}

\item{facet_by}{A character string specifying the column name of the data frame to facet the plot.}

\item{facet_scales}{A character string specifying the scales of the facets.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{keep_empty}{A logical value indicating whether to keep empty groups.
If FALSE, empty groups will be removed.}

\item{nrow}{A numeric value specifying the number of rows in the facet.}

\item{ncol}{A numeric value specifying the number of columns in the facet.}

\item{byrow}{A logical value indicating whether to fill the plots by row.}

\item{...}{Additional arguments for future expansion.}

\item{split_by}{The column(s) to split data by and plot separately or facet by.
When \code{facet} is TRUE, the columns specified here will be used to facet the plot
When \code{facet} is FALSE, the columns specified here will be used to split the data
and generate multiple plots and combine them into one
If multiple columns are specified
When \code{facet} is TRUE, up to 2 columns are allowed. If one column,
\code{ggplot2::facet_wrap} is used, the number of rows and columns is determined
by \code{nrow} and \code{ncol}
If two columns, \code{ggplot2::facet_grid} is used. The first column is used for rows
and the second column is used for columns
When \code{facet} is FALSE, a warning will be issued and the columns will be
concatenated into one column, using \code{split_by_sep} as the separator}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{combine}{Whether to combine the plots into one when facet is FALSE. Default is TRUE.}

\item{seed}{The random seed to use. Default is 8525.}
}
\value{
A ggplot object.
}
\description{
Pie chart showing the composition of the data.
}
\keyword{internal}
