% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{GSEASummaryPlot}
\alias{GSEAPlot}
\title{Objects exported from other packages}
\seealso{
\code{\link[plotthis:gsea]{plotthis::GSEASummaryPlot()}}

\code{\link[plotthis:gsea]{plotthis::GSEAPlot()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{plotthis}{\code{\link[plotthis:gsea]{GSEAPlot}}, \code{\link[plotthis:gsea]{GSEASummaryPlot}}}
}}

