% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{SpatPlot}
\alias{SpatPlot}
\alias{SpatPlot.Seurat}
\alias{SpatPlot.Seurat.Visium}
\alias{SpatPlot.Seurat.SlideSeq}
\alias{SpatPlot.Seurat.FOV}
\alias{SpatPlot.giotto}
\alias{SpatFeaturePlot}
\alias{SpatFeaturePlot.Seurat}
\alias{SpatFeaturePlot.giotto}
\alias{SpatDimPlot}
\alias{SpatDimPlot.Seurat}
\alias{SpatDimPlot.giotto}
\title{Plot features for spatial data}
\usage{
SpatPlot(object, ...)

\method{SpatPlot}{Seurat}(object, image = NULL, ...)

\method{SpatPlot}{Seurat.Visium}(
  object,
  image = NULL,
  masks = NULL,
  shapes = NULL,
  points = NULL,
  ext = NULL,
  points_x = "x",
  points_y = "y",
  image_scale = NULL,
  crop = TRUE,
  group_by = NULL,
  features = NULL,
  layer = "data",
  scale_factor = NULL,
  layers = NULL,
  flip_y = TRUE,
  theme = "theme_box",
  theme_args = list(),
  label = FALSE,
  label_size = 4,
  label_fg = "white",
  label_bg = "black",
  label_bg_r = 0.1,
  label_repel = FALSE,
  label_repulsion = 20,
  label_pt_size = 1,
  label_pt_color = "black",
  label_segment_color = "black",
  label_insitu = FALSE,
  palette = NULL,
  palette_reverse = FALSE,
  palcolor = NULL,
  highlight = NULL,
  highlight_alpha = 1,
  highlight_size = 1,
  highlight_color = "black",
  highlight_stroke = 0.8,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  ...
)

\method{SpatPlot}{Seurat.SlideSeq}(
  object,
  image = NULL,
  masks = NULL,
  shapes = NULL,
  points = NULL,
  ext = NULL,
  image_scale = NULL,
  crop = TRUE,
  group_by = NULL,
  features = NULL,
  layer = "data",
  layers = NULL,
  flip_y = TRUE,
  theme = "theme_box",
  theme_args = list(),
  label = FALSE,
  label_size = 4,
  label_fg = "white",
  label_bg = "black",
  label_bg_r = 0.1,
  label_repel = FALSE,
  label_repulsion = 20,
  label_pt_size = 1,
  label_pt_color = "black",
  label_segment_color = "black",
  label_insitu = FALSE,
  palette = NULL,
  palette_reverse = FALSE,
  palcolor = NULL,
  highlight = NULL,
  highlight_alpha = 1,
  highlight_size = 1,
  highlight_color = "black",
  highlight_stroke = 0.8,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  ...
)

\method{SpatPlot}{Seurat.FOV}(
  object,
  image = NULL,
  masks = NULL,
  shapes = NULL,
  points = NULL,
  ext = NULL,
  points_x = "x",
  points_y = "y",
  image_scale = NULL,
  crop = TRUE,
  group_by = NULL,
  features = NULL,
  layer = "data",
  scale_factor = NULL,
  layers = NULL,
  flip_y = TRUE,
  theme = "theme_box",
  theme_args = list(),
  label = FALSE,
  label_size = 4,
  label_fg = "white",
  label_bg = "black",
  label_bg_r = 0.1,
  label_repel = FALSE,
  label_repulsion = 20,
  label_pt_size = 1,
  label_pt_color = "black",
  label_segment_color = "black",
  label_insitu = FALSE,
  palette = NULL,
  palette_reverse = FALSE,
  palcolor = NULL,
  highlight = NULL,
  highlight_alpha = 1,
  highlight_size = 1,
  highlight_color = "black",
  highlight_stroke = 0.8,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  ...
)

\method{SpatPlot}{giotto}(
  object,
  image = NULL,
  masks = NULL,
  shapes = NULL,
  points = NULL,
  ext = NULL,
  layer = c("normalized", "scaled", "custom", "raw"),
  layers = NULL,
  flip_y = FALSE,
  group_by = "shapes",
  feat_type = "rna",
  features = NULL,
  nmols = 1000,
  use_overlap = FALSE,
  spat_unit = NULL,
  spat_loc_name = NULL,
  spat_enr_names = NULL,
  shapes_feat_type = "cell",
  shapes_fill_by = NULL,
  shapes_alpha = NULL,
  theme = "theme_box",
  theme_args = list(),
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  ...
)

SpatFeaturePlot(object, ...)

\method{SpatFeaturePlot}{Seurat}(object, image = NULL, ...)

\method{SpatFeaturePlot}{giotto}(object, image = NULL, group_by = NULL, ...)

SpatDimPlot(object, ...)

\method{SpatDimPlot}{Seurat}(object, image = NULL, group_by = NULL, ...)

\method{SpatDimPlot}{giotto}(object, image = NULL, group_by = NULL, ...)
}
\arguments{
\item{object}{A Seurat object or a Giotto object.}

\item{...}{Additional arguments passed to the plotting functions.}

\item{image}{The name of the image to plot. If NULL, the first image in the object will be used.}
}
\value{
A ggplot object

A ggplot object

A ggplot object
}
\description{
The features can include  expression, dimension reduction components, metadata, etc

The features can include  expression, dimension reduction components, metadata, etc
}
\keyword{internal}
