% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_seurat}
\alias{subset_seurat}
\title{Subset function with automatic layer repair for Seurat v5 objects}
\usage{
subset_seurat(object, ...)
}
\arguments{
\item{object}{Seurat object to subset}

\item{...}{Arguments passed to \code{subset()}}
}
\value{
Subsetted Seurat object with repaired layers if needed
}
\description{
This function handles corrupted layers in Seurat v5 assays that can cause
"incorrect number of dimensions" errors during subsetting.
}
\details{
Strategy:
\enumerate{
\item Evaluate subset normally
\item Check for corrupted layers proactively
\item Repair any corrupted layers before subsetting
\item Perform the subset operation
\item If subset still fails, re-throw the error
}

Example: subset_seurat(obj, EFS \%in\% c("EFS_L", "EFS_S"), nCount_RNA > 1000)
}
\keyword{internal}
