% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{theme_scp}
\alias{theme_scp}
\title{A ggplot2 theme and palettes for scplotter
Borrowed from the \code{theme_scp} function in the \code{SCP} pipeline}
\usage{
theme_scp(aspect.ratio = NULL, base_size = 12, font_family = NULL, ...)
}
\arguments{
\item{aspect.ratio}{The aspect ratio of the plot}

\item{base_size}{The base size of the text}

\item{font_family}{The font family of the text}

\item{...}{Other arguments for \code{theme()}}
}
\value{
A ggplot2 theme
}
\description{
A ggplot2 theme and palettes for scplotter
Borrowed from the \code{theme_scp} function in the \code{SCP} pipeline
}
\seealso{
\url{https://github.com/zhanghao-njmu/SCP}
}
