% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_common_arguments}
\alias{validate_common_arguments}
\title{Common params for plotting functions and validate the ones related to plotting}
\usage{
validate_common_arguments(
  split_by,
  seed,
  facet,
  plot_type = NULL,
  split_by_sep = "_",
  group_by = NULL,
  group_by_sep = "_",
  facet_scales = "fixed",
  theme = "theme_scp",
  theme_args = list(),
  palette = NULL,
  palcolor = NULL,
  keep_empty = FALSE,
  alpha = 1,
  x_text_angle = 0,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  ...
)
}
\arguments{
\item{split_by}{The column(s) to split data by and plot separately or facet by.
When \code{facet} is TRUE, the columns specified here will be used to facet the plot
When \code{facet} is FALSE, the columns specified here will be used to split the data
and generate multiple plots and combine them into one
If multiple columns are specified
When \code{facet} is TRUE, up to 2 columns are allowed. If one column,
\code{ggplot2::facet_wrap} is used, the number of rows and columns is determined
by \code{nrow} and \code{ncol}
If two columns, \code{ggplot2::facet_grid} is used. The first column is used for rows
and the second column is used for columns
When \code{facet} is FALSE, a warning will be issued and the columns will be
concatenated into one column, using \code{split_by_sep} as the separator}

\item{seed}{The random seed to use. Default is 8525.}

\item{facet}{Whether to facet the plot. Default is FALSE.
Otherwise, the data will be split by \code{\link{split_by}} and generate multiple plots
and combine them into one using \code{patchwork::wrap_plots}}

\item{plot_type}{The type of plot.}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{group_by}{Columns to group the data for plotting
For those plotting functions that do not support multiple groups,
They will be concatenated into one column, using \code{\link{group_by_sep}} as the separator}

\item{group_by_sep}{The separator for multiple group_by columns. See \code{\link{group_by}}}

\item{facet_scales}{Whether to scale the axes of facets. Default is "fixed"
Other options are "free", "free_x", "free_y". See \code{\link{ggplot2::facet_wrap}}}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{keep_empty}{A logical value indicating whether to keep empty groups.
If FALSE, empty groups will be removed.}

\item{alpha}{A numeric value specifying the transparency of the plot.}

\item{x_text_angle}{A numeric value specifying the angle of the x-axis text.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{combine}{Whether to combine the plots into one when facet is FALSE. Default is TRUE.}

\item{nrow}{The number of rows in facet_wrap/wrap_plots. Default is NULL.}

\item{ncol}{The number of columns in facet_wrap/wrap_plots. Default is NULL.}

\item{byrow}{Whether to fill the plots by row. Default is TRUE.}

\item{...}{Additional arguments for future expansion.}
}
\description{
Common params for plotting functions and validate the ones related to plotting
}
\keyword{internal}
