% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccplot.R
\name{CCCPlot}
\alias{CCCPlot}
\title{Cell-Cell Communication Plot}
\usage{
CCCPlot(
  data,
  plot_type = c("dot", "network", "chord", "circos", "heatmap", "sankey", "alluvial"),
  method = c("aggregation", "interaction"),
  magnitude = waiver(),
  specificity = waiver(),
  weighted = TRUE,
  meta_specificity = "sumlog",
  split_by = NULL,
  x_text_angle = 90,
  link_curvature = 0.2,
  link_alpha = 0.6,
  facet_by = NULL,
  show_row_names = TRUE,
  show_column_names = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame with the cell-cell communication data.
A typical data frame should have the following columns:
\itemize{
\item \code{source} The source cell type.
\item \code{target} The target cell type.
\item \code{ligand} The ligand gene.
\item \code{receptor} The receptor gene.
\item \code{ligand_means} The mean expression of the ligand gene per cell type.
\item \code{receptor_means} The mean expression of the receptor gene per cell type.
\item \code{ligand_props} The proportion of cells that express the entity.
\item \code{receptor_props} The proportion of cells that express the entity.
\item \verb{<magnitude>} The magnitude of the communication.
\item \verb{<specificity>} The specificity of the communication.
Depends on the \code{plot_type}, some columns are optional. But the \code{source}, \code{target},
\code{ligand}, \code{receptor} and \verb{<magnitude>} are required.
}}

\item{plot_type}{The type of plot to use. Default is "dot".
Possible values are "network", "chord", "circos", "heatmap", "sankey", "alluvial", and "dot".
\itemize{
\item network: A network plot with the source and target cells as the nodes and the communication as the edges.
\item chord: A chord plot with the source and target cells as the nodes and the communication as the chords.
\item circos: Alias of "chord".
\item heatmap: A heatmap plot with the source and target cells as the rows and columns.
\item sankey: A sankey plot with the source and target cells as the nodes and the communication as the flows.
\item alluvial: Alias of "sankey".
\item dot: A dot plot with the source and target cells as the nodes and the communication as the dots.
}}

\item{method}{The method to determine the plot entities.
\itemize{
\item aggregation: Aggregate the ligand-receptor pairs interactions for each source-target pair.
Only the source / target pairs will be plotted.
\item interaction: Plot the ligand-receptor pairs interactions directly.
The ligand-receptor pairs will also be plotted.
}}

\item{magnitude}{The column name in the data to use as the magnitude of the communication.
By default, the second last column will be used.
See \code{li.mt.show_methods()} for the available methods in \code{LIANA}.
or \url{https://liana-py.readthedocs.io/en/latest/notebooks/basic_usage.html#Tileplot}}

\item{specificity}{The column name in the data to use as the specificity of the communication.
By default, the last column will be used.
If the method doesn't have a specificity, set it to NULL.}

\item{weighted}{Whether to use the magnitude as the weight of the aggregation interaction strength.
for source-target pairs. Default is TRUE. Otherwise, the number of interactions will be used.
Only used when \code{method} is "aggregation".}

\item{meta_specificity}{The method to calculate the specificity when there are multiple
ligand-receptor pairs interactions. Default is "sumlog".
It should be one of the methods in the \code{metap} package.}

\item{split_by}{A character vector of column names to split the plots. Default is NULL.}

\item{x_text_angle}{The angle of the x-axis text. Default is 90.
Only used when \code{plot_type} is "dot".}

\item{link_curvature}{The curvature of the links. Default is 0.2.
Only used when \code{plot_type} is "network".}

\item{link_alpha}{The transparency of the links. Default is 0.6.
Only used when \code{plot_type} is "network".}

\item{facet_by}{A character vector of column names to facet the plots. Default is NULL.
It should always be NULL.}

\item{show_row_names}{Whether to show the row names in the heatmap. Default is TRUE.
Only used when \code{plot_type} is "heatmap".}

\item{show_column_names}{Whether to show the column names in the heatmap. Default is TRUE.
Only used when \code{plot_type} is "heatmap".}

\item{...}{Other arguments passed to the specific plot function.
\itemize{
\item For \code{Network}, see \link[plotthis:Network]{plotthis::Network}.
\item For \code{ChordPlot}, see \link[plotthis:chordplot]{plotthis::ChordPlot}.
\item For \code{Heatmap}, see \link[plotthis:Heatmap]{plotthis::Heatmap}.
\item For \code{SankeyPlot}, see \link[plotthis:sankeyplot]{plotthis::SankeyPlot}.
\item For \code{DotPlot}, see \link[plotthis:dotplot]{plotthis::DotPlot}.
}}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Plot the cell-cell communication.
See also:
\itemize{
\item The review: \url{https://www.sciencedirect.com/science/article/pii/S2452310021000081}
\item The LIANA package: \url{https://liana-py.readthedocs.io/en/latest/notebooks/basic_usage.html#Tileplot}
\item The CCPlotR package: \url{https://github.com/Sarah145/CCPlotR}
}
}
\examples{
set.seed(8525)
data(cellphonedb_res)
CCCPlot(data = cellphonedb_res, plot_type = "network", legend.position = "none",
 theme = "theme_blank", theme_args = list(add_coord = FALSE))
CCCPlot(cellphonedb_res, plot_type = "chord")
CCCPlot(cellphonedb_res, plot_type = "heatmap")
CCCPlot(cellphonedb_res, plot_type = "dot", weighted = FALSE)

cellphonedb_res_sub <- cellphonedb_res[
  cellphonedb_res$source \%in\% c("Dendritic", "CD14+ Monocyte"),]
CCCPlot(cellphonedb_res_sub, plot_type = "dot", method = "interaction")
CCCPlot(cellphonedb_res_sub, plot_type = "network", method = "interaction",
  node_size_by = 1)
CCCPlot(cellphonedb_res_sub, plot_type = "heatmap", method = "interaction")
}
