% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldimplot.R
\name{CellDimPlot}
\alias{CellDimPlot}
\title{Cell Dimension Reduction Plot}
\usage{
CellDimPlot(
  object,
  reduction = NULL,
  graph = NULL,
  group_by = NULL,
  ident = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  velocity = NULL,
  ...
)
}
\arguments{
\item{object}{A seurat object, a giotto object, a path to an .h5ad file or an opened \code{H5File} by \code{hdf5r} package.}

\item{reduction}{Name of the reduction to plot (for example, "umap").}

\item{graph}{Specify the graph name to add edges between cell neighbors to the plot.}

\item{group_by}{A character vector of column name(s) to group the data. Default is NULL.}

\item{ident}{Alias for \code{group_by}.}

\item{spat_unit}{The spatial unit to use for the plot. Only applied to Giotto objects.}

\item{feat_type}{feature type of the features (e.g. "rna", "dna", "protein"), only applied to Giotto objects.}

\item{velocity}{The name of velocity reduction to plot cell velocities.
It is typically \code{"stochastic_<reduction>"}, \code{"deterministic_<reduction>"}, or \code{"dynamical_<reduction>"}.}

\item{...}{Other arguments passed to \code{\link[plotthis:dimplot]{plotthis::DimPlot()}}.}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
This function creates a dimension reduction plot for a Seurat object
a Giotto object, a path to an .h5ad file or an opened \code{H5File} by \code{hdf5r} package.
It allows for various customizations such as grouping by metadata,
adding edges between cell neighbors, highlighting specific cells, and more.
This function is a wrapper around \code{\link[plotthis:dimplot]{plotthis::DimPlot()}}, which provides a
flexible way to visualize cell clusters in reduced dimensions. This function
extracts the necessary data from the Seurat or Giotto object and passes it to
\code{\link[plotthis:dimplot]{plotthis::DimPlot()}}.
}
\details{
See
\itemize{
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_CODEX.html}
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_seqFISH.html}
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_SlideSeq.html}
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_Spatial_CITE-Seq.html}
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_Visium.html}
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_VisiumHD.html}
\item \url{https://pwwang.github.io/scplotter/articles/Giotto_Xenium.html}
}

for examples of using this function with a Giotto object.

And see:
\itemize{
\item \url{https://pwwang.github.io/scplotter/articles/Working_with_anndata_h5ad_files.html}
}

for examples of using this function with .h5ad files.
}
\examples{
\donttest{
set.seed(8525)
data(pancreas_sub)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP")
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            theme = "theme_blank")
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            palette = "seurat", theme = "theme_blank")
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            theme = ggplot2::theme_classic, theme_args = list(base_size = 16))
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            raster = TRUE, raster_dpi = 30)

# Highlight cells
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", reduction = "UMAP",
  highlight = 'SubCellType == "Epsilon"'
)
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", split_by = "Phase", reduction = "UMAP",
  highlight = TRUE, theme = "theme_blank", legend.position = "none"
)
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", facet_by = "Phase", reduction = "UMAP",
  highlight = TRUE, theme = "theme_blank", legend.position = "none"
)

# Add group labels
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            label = TRUE)
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", reduction = "UMAP",
  label = TRUE, label_fg = "orange", label_bg = "red", label_size = 5
)
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", reduction = "UMAP",
  label = TRUE, label_insitu = TRUE
)
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", reduction = "UMAP",
  label = TRUE, label_insitu = TRUE, label_repel = TRUE,
  label_segment_color = "red"
)

# Add various shape of marks
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE, mark_expand = grid::unit(1, "mm"))
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE, mark_alpha = 0.3)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE, mark_linetype = 2)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE, mark_type = "ellipse")
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE, mark_type = "rect")
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_mark = TRUE, mark_type = "circle")

# Add a density layer
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_density = TRUE)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            add_density = TRUE, density_filled = TRUE)
CellDimPlot(pancreas_sub,
  group_by = "SubCellType", reduction = "UMAP",
  add_density = TRUE, density_filled = TRUE, density_filled_palette = "Blues",
  highlight = TRUE
)

# Add statistical charts
CellDimPlot(pancreas_sub,
  group_by = "CellType", reduction = "UMAP", stat_by = "Phase")
CellDimPlot(pancreas_sub,
  group_by = "CellType", reduction = "UMAP", stat_by = "Phase",
  stat_plot_type = "ring", stat_plot_label = TRUE, stat_plot_size = 0.15)
CellDimPlot(pancreas_sub,
  group_by = "CellType", reduction = "UMAP", stat_by = "Phase",
  stat_plot_type = "bar", stat_type = "count")
CellDimPlot(pancreas_sub,
  group_by = "CellType", reduction = "UMAP", stat_by = "Phase",
  stat_plot_type = "line", stat_type = "count", stat_args = list(point_size = 1))

# Chane the plot type from point to the hexagonal bin
CellDimPlot(pancreas_sub, group_by = "CellType", reduction = "UMAP",
            hex = TRUE)
CellDimPlot(pancreas_sub, group_by = "CellType", reduction = "UMAP",
            hex = TRUE, hex_bins = 20)
CellDimPlot(pancreas_sub, group_by = "CellType", reduction = "UMAP",
            hex = TRUE, hex_count = FALSE)

# Show neighbors graphs on the plot
CellDimPlot(pancreas_sub, group_by = "CellType", reduction = "UMAP",
            graph = "RNA_nn")
CellDimPlot(pancreas_sub, group_by = "CellType", reduction = "UMAP",
            graph = "RNA_snn", edge_color = "grey80")

# Show lineages on the plot based on the pseudotime
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            lineages = paste0("Lineage", 1:3))
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            lineages = paste0("Lineage", 1:3), lineages_whiskers = TRUE)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            lineages = paste0("Lineage", 1:3), lineages_span = 0.1)

# Velocity
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "PCA",
  velocity = "stochastic_PCA")
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "PCA",
  velocity = "stochastic_PCA", velocity_plot_type = "grid", pt_alpha = 0.5)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "PCA",
  velocity = "stochastic_PCA", velocity_plot_type = "stream", pt_alpha = 0.5)

# 3D plot
pancreas_sub@reductions$UMAP@cell.embeddings <- cbind(
 pancreas_sub@reductions$UMAP@cell.embeddings,
 umap_3 = rnorm(1000)  # fake the 3rd dimension
)
CellDimPlot(pancreas_sub, group_by = "SubCellType", reduction = "UMAP",
            dims = 1:3, label = TRUE)
}
}
\seealso{
\code{\link[=CellStatPlot]{CellStatPlot()}} \code{\link[=CellVelocityPlot]{CellVelocityPlot()}}
}
