% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldimplot.R
\name{CellVelocityPlot}
\alias{CellVelocityPlot}
\title{Cell Velocity Plot}
\usage{
CellVelocityPlot(
  object,
  reduction,
  v_reduction,
  spat_unit = NULL,
  feat_type = NULL,
  ...
)
}
\arguments{
\item{object}{A seurat object or a giotto object.}

\item{reduction}{Name of the reduction to plot (for example, "umap").}

\item{v_reduction}{Name of the velocity reduction to plot (for example, "stochastic_umap").
It should be the same as the reduction used to calculate the velocity.}

\item{spat_unit}{The spatial unit to use for the plot. Only applied to Giotto objects.}

\item{feat_type}{feature type of the features (e.g. "rna", "dna", "protein"), only applied to Giotto objects.}

\item{...}{Other arguments passed to \code{\link[plotthis:VelocityPlot]{plotthis::VelocityPlot()}}.}
}
\value{
A ggplot object
}
\description{
This function creates a cell velocity plot for a Seurat object
or a Giotto object. It allows for various customizations such as grouping by metadata,
adding edges between cell neighbors, highlighting specific cells, and more.
This function is a wrapper around \code{\link[plotthis:VelocityPlot]{plotthis::VelocityPlot()}}, which provides a
flexible way to visualize cell velocities in reduced dimensions. This function
extracts the cell embeddings and velocity embeddings from the Seurat or Giotto object
and passes them to \code{\link[plotthis:VelocityPlot]{plotthis::VelocityPlot()}}.
}
\seealso{
\code{\link[=CellDimPlot]{CellDimPlot()}}
}
