% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldimplot.R
\name{CellVelocityPlot}
\alias{CellVelocityPlot}
\title{Cell Velocity Plot}
\usage{
CellVelocityPlot(
  object,
  reduction,
  v_reduction,
  spat_unit = NULL,
  feat_type = NULL,
  group_by = NULL,
  ...
)
}
\arguments{
\item{object}{A seurat object, a giotto object, a path to an .h5ad file or an opened \code{H5File} by \code{hdf5r} package.}

\item{reduction}{Name of the reduction to plot (for example, "umap").}

\item{v_reduction}{Name of the velocity reduction to plot (for example, "stochastic_umap").
It should be the same as the reduction used to calculate the velocity.}

\item{spat_unit}{The spatial unit to use for the plot. Only applied to Giotto objects.}

\item{feat_type}{feature type of the features (e.g. "rna", "dna", "protein"), only applied to Giotto objects.}

\item{group_by}{A character vector of metadata column name(s) to group (color) the data. Default is NULL.}

\item{...}{Other arguments passed to \code{\link[plotthis:VelocityPlot]{plotthis::VelocityPlot()}}.}
}
\value{
A ggplot object
}
\description{
This function creates a cell velocity plot for a Seurat object,
a Giotto object, a path to an .h5ad file or an opened \code{H5File} by \code{hdf5r} package.
It allows for various customizations such as grouping by metadata,
adding edges between cell neighbors, highlighting specific cells, and more.
This function is a wrapper around \code{\link[plotthis:VelocityPlot]{plotthis::VelocityPlot()}}, which provides a
flexible way to visualize cell velocities in reduced dimensions. This function
extracts the cell embeddings and velocity embeddings from the Seurat or Giotto object
and passes them to \code{\link[plotthis:VelocityPlot]{plotthis::VelocityPlot()}}.
}
\details{
See:
\itemize{
\item \url{https://pwwang.github.io/scplotter/articles/Working_with_anndata_h5ad_files.html}
}

for examples of using this function with .h5ad files.
}
\examples{
\donttest{
data(pancreas_sub)

CellVelocityPlot(pancreas_sub, reduction = "PCA", v_reduction = "stochastic_PCA")
CellVelocityPlot(pancreas_sub, reduction = "PCA", v_reduction = "stochastic_PCA",
 plot_type = "grid")
CellVelocityPlot(pancreas_sub, reduction = "PCA", v_reduction = "stochastic_PCA",
 plot_type = "stream")
CellVelocityPlot(pancreas_sub, reduction = "PCA", v_reduction = "stochastic_PCA",
 group_by = "SubCellType")
}
}
\seealso{
\code{\link[=CellDimPlot]{CellDimPlot()}}
}
