% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonaldivplot.R
\name{ClonalDiversityPlot}
\alias{ClonalDiversityPlot}
\title{ClonalDiversityPlot}
\usage{
ClonalDiversityPlot(
  data,
  clone_call = "gene",
  chain = "both",
  method = c("shannon", "gini.coeff", "inv.simpson", "norm.entropy", "gini.simpson",
    "chao1", "ACE"),
  plot_type = c("bar", "box", "violin"),
  position = "dodge",
  group_by = NULL,
  facet_by = NULL,
  split_by = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or
\link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{clone_call}{How to call the clone - VDJC gene (gene), CDR3 nucleotide (nt),
CDR3 amino acid (aa), VDJC gene + CDR3 nucleotide (strict) or a custom variable
in the data}

\item{chain}{indicate if both or a specific chain should be used - e.g. "both",
"TRA", "TRG", "IGH", "IGL"}

\item{method}{The method to calculate the diversity. Options are "shannon" (default),
"inv.simpson", "norm.entropy", "gini.simpson", "chao1" and "ACE".
See \link[scRepertoire:clonalDiversity]{scRepertoire::clonalDiversity} for details.}

\item{plot_type}{The type of plot. Options are "bar", "box" and "violin".}

\item{position}{The position adjustment for the bars. Default is "dodge".}

\item{group_by}{A character vector of column names to group the samples. Default is NULL.}

\item{facet_by}{A character vector of column names to facet the plots. Default is NULL.}

\item{split_by}{A character vector of column names to split the plots. Default is NULL.}

\item{xlab}{The x-axis label. Default is NULL.}

\item{ylab}{The y-axis label. Default is NULL.}

\item{...}{Other arguments passed to the specific plot function.
\itemize{
\item For "bar", \link[plotthis:barplot]{plotthis::BarPlot}
\item For "box", \link[plotthis:boxviolinplot]{plotthis::BoxPlot}
\item For "violin", \link[plotthis:boxviolinplot]{plotthis::ViolinPlot}
}}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Plot the clonal diversities of the samples/groups.
}
\examples{
set.seed(8525)
data(contig_list, package = "scRepertoire")
data <- scRepertoire::combineTCR(contig_list,
    samples = c("P17B", "P17L", "P18B", "P18L", "P19B","P19L", "P20B", "P20L"))
data <- scRepertoire::addVariable(data,
    variable.name = "Type",
    variables = rep(c("B", "L"), 4)
)
data <- scRepertoire::addVariable(data,
    variable.name = "Subject",
    variables = rep(c("P17", "P18", "P19", "P20"), each = 2)
)

ClonalDiversityPlot(data)
ClonalDiversityPlot(data, group_by = "Type")
ClonalDiversityPlot(data, group_by = "Type", plot_type = "box")
ClonalDiversityPlot(data, group_by = "Type", plot_type = "violin")
ClonalDiversityPlot(data, group_by = "Type", plot_type = "violin",
  method = "gini.coeff", add_box = TRUE)
}
