% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalpositionalplot.R
\name{ClonalKmerPlot}
\alias{ClonalKmerPlot}
\title{ClonalKmerPlot}
\usage{
ClonalKmerPlot(
  data,
  chain = "TRB",
  clone_call = "aa",
  k = 3,
  top = 25,
  group_by = "Sample",
  group_by_sep = "_",
  facet_by = NULL,
  split_by = NULL,
  plot_type = c("bar", "line", "heatmap"),
  theme_args = list(),
  aspect.ratio = NULL,
  facet_ncol = NULL,
  ...
)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or
\link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{chain}{The chain to be analyzed. Default is "TRB".}

\item{clone_call}{The column name of the clone call. Default is "aa".}

\item{k}{The length of the k-mer. Default is 3.}

\item{top}{The number of top k-mers to display. Default is 25.}

\item{group_by}{The variable to group the data by. Default is "Sample".}

\item{group_by_sep}{The separator to use when combining groupings. Default is "_".}

\item{facet_by}{A character vector of column names to facet the plots. Default is NULL.}

\item{split_by}{A character vector of column names to split the plots. Default is NULL.}

\item{plot_type}{The type of plot to generate. Default is "bar".
\itemize{
\item "bar": Bar plot.
\item "line": Line plot.
\item "heatmap": Heatmap.
}}

\item{theme_args}{A list of arguments to be passed to the \link[ggplot2:theme]{ggplot2::theme} function.}

\item{aspect.ratio}{The aspect ratio of the plot. Default is NULL.}

\item{facet_ncol}{The number of columns in the facet grid. Default is NULL.}

\item{...}{Other arguments passed to the specific plot function.
\itemize{
\item For "bar", \code{\link[plotthis:barplot]{plotthis::BarPlot()}}.
\item For "line", \code{\link[plotthis:LinePlot]{plotthis::LinePlot()}}.
\item For "heatmap", \code{\link[plotthis:Heatmap]{plotthis::Heatmap()}}.
}}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Explore the k-mer frequency of CDR3 sequences.
}
\examples{
set.seed(8525)
data(contig_list, package = "scRepertoire")
data <- scRepertoire::combineTCR(contig_list,
    samples = c("P17B", "P17L", "P18B", "P18L", "P19B","P19L", "P20B", "P20L"))
data <- scRepertoire::addVariable(data,
    variable.name = "Type",
    variables = rep(c("B", "L"), 4)
)
data <- scRepertoire::addVariable(data,
    variable.name = "Subject",
    variables = rep(c("P17", "P18", "P19", "P20"), each = 2)
)

ClonalKmerPlot(data)
ClonalKmerPlot(data, group_by = "Type")
ClonalKmerPlot(data, group_by = "Type", plot_type = "line")
ClonalKmerPlot(data, group_by = "Type", plot_type = "heatmap")
}
