% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalstatplot.R
\name{ClonalLengthPlot}
\alias{ClonalLengthPlot}
\title{ClonalLengthPlot}
\usage{
ClonalLengthPlot(
  data,
  clone_call = "aa",
  chain = "both",
  plot_type = c("bar", "box", "violin", "density"),
  x_nbreaks = 10,
  group_by = "Sample",
  order = list(),
  xlab = "Length",
  ylab = NULL,
  position = "dodge",
  facet_by = NULL,
  split_by = NULL,
  ...
)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or
\link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{clone_call}{How to call the clone - only "nt" or "aa" is supported.}

\item{chain}{indicate if both or a specific chain should be used - e.g. "both",
"TRA", "TRB", "TRD", "TRG", "IGH", or "IGL" to specify a specific chain.}

\item{plot_type}{The type of plot to use. Default is "bar".
Possible values are "box", "violin" and "density".}

\item{x_nbreaks}{The number of breaks for the x-axis. Default is 10.}

\item{group_by}{The column name in the meta data to group the cells. Default: "Sample"}

\item{order}{The order of the groups. Default is an empty list.
It should be a list of values. The names are the column names, and the values are the order.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{position}{The position of the bars for bar plot on the x-axis. Default is "dodge".}

\item{facet_by}{The column name in the meta data to facet the plots. Default: NULL}

\item{split_by}{The column name in the meta data to split the plots. Default: NULL}

\item{...}{Other arguments passed to the specific plot function.
\itemize{
\item For \code{bar} plot, see \code{\link[plotthis:barplot]{plotthis::BarPlot()}}.
\item For \code{box} plot, see \code{\link[plotthis:boxviolinplot]{plotthis::BoxPlot()}}.
\item For \code{violin} plot, see \code{\link[plotthis:boxviolinplot]{plotthis::ViolinPlot()}}.
\item For \code{density} plot, see \code{\link[plotthis:densityhistoplot]{plotthis::DensityPlot()}}.
}}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Plot the length distribution of the CDR3 sequences
}
\examples{
set.seed(8525)
data(contig_list, package = "scRepertoire")
data <- scRepertoire::combineTCR(contig_list)
data <- scRepertoire::addVariable(data, variable.name = "Type",
 variables = sample(c("B", "L"), 8, replace = TRUE))
data <- scRepertoire::addVariable(data, variable.name = "Sex",
 variables = sample(c("M", "F"), 8, replace = TRUE))

ClonalLengthPlot(data)
ClonalLengthPlot(data, plot_type = "box")
ClonalLengthPlot(data, clone_call = "nt", plot_type = "violin", chain = "TRB",
 group_by = "Type", comparisons = TRUE)
ClonalLengthPlot(data, plot_type = "density", chain = "TRA")
}
