% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalstatplot.R
\name{ClonalOverlapPlot}
\alias{ClonalOverlapPlot}
\title{ClonalOverlapPlot}
\usage{
ClonalOverlapPlot(
  data,
  clone_call = "aa",
  chain = "both",
  group_by = "Sample",
  group_by_sep = "_",
  full = TRUE,
  split_by = NULL,
  order = list(),
  method = c("raw", "overlap", "morisita", "jaccard", "cosine"),
  palette = "Blues",
  label_accuracy = NULL,
  label_cutoff = 0.001,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  ...
)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or
\link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{clone_call}{How to call the clone - VDJC gene (gene), CDR3 nucleotide (nt),
CDR3 amino acid (aa), VDJC gene + CDR3 nucleotide (strict) or a custom variable
in the data}

\item{chain}{indicate if both or a specific chain should be used - e.g. "both",
"TRA", "TRG", "IGH", "IGL"}

\item{group_by}{The column name in the meta data to group the cells. Default: "Sample"}

\item{group_by_sep}{The separator used to concatenate the group_by when multiple columns are used.}

\item{full}{Whether to plot the full heatmap, or just a triangle. Default is TRUE.}

\item{split_by}{The column name in the meta data to split the plots. Default: NULL}

\item{order}{The order of the groups. Default is an empty list.
It should be a list of values. The names are the column names, and the values are the order.}

\item{method}{The method to calculate the overlap. Default is "raw".
\itemize{
\item "overlap" - overlap coefficient
\item "morisita" - Morisita’s overlap index
\item "jaccard" - Jaccard index
\item "cosine" - cosine similarity
\item "raw" - exact number of overlapping clones
See also \link[scRepertoire:clonalOverlap]{scRepertoire::clonalOverlap}.
}}

\item{palette}{The color palette to use. Default is "Blues".}

\item{label_accuracy}{The accuracy of the labels. Default is NULL.
If NULL, it will be 1 for "raw" and 0.01 for other methods.}

\item{label_cutoff}{The cutoff for the labels to show. Default is 1e-3.}

\item{cluster_rows}{Whether to cluster the rows. Default is FALSE.}

\item{cluster_columns}{Whether to cluster the columns. Default is FALSE.}

\item{show_row_names}{Whether to show the row names. Default is TRUE.}

\item{show_column_names}{Whether to show the column names. Default is TRUE.}

\item{...}{Other arguments passed to the specific plot function \code{\link[plotthis:Heatmap]{plotthis::Heatmap()}}.}
}
\value{
A ComplexHeatmap object or a list if \code{combine} is FALSE
}
\description{
Plot the overlap of the clones in different samples/groups.
}
\examples{
set.seed(8525)
data(contig_list, package = "scRepertoire")
data <- scRepertoire::combineTCR(contig_list,
    samples = c("P17B", "P17L", "P18B", "P18L", "P19B","P19L", "P20B", "P20L"))
data <- scRepertoire::addVariable(data,
    variable.name = "Type",
    variables = rep(c("B", "L"), 4)
)
data <- scRepertoire::addVariable(data,
    variable.name = "Subject",
    variables = rep(c("P17", "P18", "P19", "P20"), each = 2)
)

ClonalOverlapPlot(data)
ClonalOverlapPlot(data, clone_call = "strict", label_cutoff = 0,
  label_accuracy = 0.001, method = "morisita", full = FALSE)
ClonalOverlapPlot(data, group_by = c("Subject", "Type"))
ClonalOverlapPlot(data, group_by = "Type", split_by = "Subject")
}
