% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonaldivplot.R
\name{ClonalRarefactionPlot}
\alias{ClonalRarefactionPlot}
\title{ClonalRarefactionPlot}
\usage{
ClonalRarefactionPlot(
  data,
  clone_call = "aa",
  chain = "both",
  group_by = "Sample",
  group_by_sep = "_",
  n_boots = 20,
  q = 0,
  facet_by = NULL,
  split_by = NULL,
  split_by_sep = "_",
  palette = "Paired",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  ...
)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or
\link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{clone_call}{How to call the clone - VDJC gene (gene), CDR3 nucleotide (nt),
CDR3 amino acid (aa), VDJC gene + CDR3 nucleotide (strict) or a custom variable}

\item{chain}{indicate if both or a specific chain should be used - e.g. "both",
"TRA", "TRG", "IGH", "IGL"}

\item{group_by}{A character vector of column names to group the samples. Default is "Sample".}

\item{group_by_sep}{The separator for the group_by column. Default is "_".}

\item{n_boots}{The number of bootstrap samples. Default is 20.}

\item{q}{The hill number. Default is 0.
\itemize{
\item 0 - Species richness
\item 1 - Shannon entropy
\item 2 - Simpson index#'
}}

\item{facet_by}{A character vector of column names to facet the plots. Default is NULL.}

\item{split_by}{A character vector of column names to split the plots. Default is NULL.}

\item{split_by_sep}{The separator for the split_by column. Default is "_".}

\item{palette}{The color palette to use. Default is "Paired".}

\item{combine}{Whether to combine the plots into a single plot. Default is TRUE.}

\item{nrow}{The number of rows in the combined plot. Default is NULL.}

\item{ncol}{The number of columns in the combined plot. Default is NULL.}

\item{byrow}{Whether to fill the combined plot by row. Default is TRUE.}

\item{...}{Other arguments passed to \code{\link[plotthis:RarefactionPlot]{plotthis::RarefactionPlot()}}.}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Plot the rarefaction curves
}
\examples{
set.seed(8525)
data(contig_list, package = "scRepertoire")
data <- scRepertoire::combineTCR(contig_list,
    samples = c("P17B", "P17L", "P18B", "P18L", "P19B","P19L", "P20B", "P20L"))
data <- scRepertoire::addVariable(data,
    variable.name = "Type",
    variables = rep(c("B", "L"), 4)
)
data <- scRepertoire::addVariable(data,
    variable.name = "Subject",
    variables = rep(c("P17", "P18", "P19", "P20"), each = 2)
)

ClonalRarefactionPlot(data, type = 1, q = 0, n_boots = 2)
ClonalRarefactionPlot(data, type = 2, q = 0, n_boots = 2)
ClonalRarefactionPlot(data, type = 3, q = 0, n_boots = 2)
ClonalRarefactionPlot(data, q = 1, n_boots = 2)
ClonalRarefactionPlot(data, q = 1, n_boots = 2, group_by = "Type")
ClonalRarefactionPlot(data, n_boots = 2, split_by = "Type")
}
