% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustreeplot.R
\name{ClustreePlot}
\alias{ClustreePlot}
\title{Clustree plot}
\usage{
ClustreePlot(object, ...)
}
\arguments{
\item{object}{The data frame or Seurat object}

\item{...}{Other arguments passed to \link[plotthis:ClustreePlot]{plotthis::ClustreePlot}}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
This function generates a clustree plot from a data frame or a Seurat object.
}
\examples{
data(ifnb_sub)
ClustreePlot(ifnb_sub, prefix = "RNA_snn_res.")
}
