% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_plot.R
\name{ClustreePlot}
\alias{ClustreePlot}
\alias{ClustreePlot.default}
\alias{ClustreePlot.data.frame}
\alias{ClustreePlot.Seurat}
\title{Clustree plot}
\usage{
ClustreePlot(x, prefix, ...)

\method{ClustreePlot}{default}(x, prefix, ...)

\method{ClustreePlot}{data.frame}(
  x,
  prefix,
  palette = "Paired",
  palcolor = NULL,
  edge_palette = "YlOrRd",
  edge_palcolor = NULL,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  theme = "theme_scp",
  theme_args = list(),
  res = 100,
  guess_size = FALSE,
  seed = 8525,
  ...
)

\method{ClustreePlot}{Seurat}(x, prefix, ...)
}
\arguments{
\item{x}{The data frame or Seurat object}

\item{prefix}{The prefix of cluster columns}

\item{...}{Additional arguments to be passed to the clustree::clustree() function}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{edge_palette}{The color palette to use for edges. Default is "YlOrRd".}

\item{edge_palcolor}{Custom colors for edge_palette. Default is NULL.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_scp".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{seed}{The random seed to use. Default is 8525.}
}
\value{
A ggplot object or a list with the plot and the height and width of the plot if guess_size is TRUE
}
\description{
This function generates a clustree plot from a data frame or a Seurat object.
}
\examples{
data(ifnb_sub)
ClustreePlot(ifnb_sub, prefix = "RNA_snn_res.")
}
