% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_call}
\alias{do_call}
\title{Call a function with a list of arguments}
\usage{
do_call(fn, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{fn}{A function to call}

\item{args}{A list of arguments to pass to the function}

\item{quote}{Whether to quote the arguments}

\item{envir}{The environment to evaluate the function in}
}
\value{
The result of the function call
}
\description{
Different from \code{do.call}, this is a faster version, especially when there are
big objects in the list.
}
\keyword{internal}
