% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurestatplot.R
\name{.feature_stat_plot}
\alias{.feature_stat_plot}
\title{Feature statistic plot when given a composed data frame}
\usage{
.feature_stat_plot(
  data,
  features,
  plot_type,
  should_shrink,
  should_pivot,
  graph = NULL,
  bg_cutoff = 0,
  dims = 1:2,
  rows_name = "Features",
  ident = NULL,
  agg = mean,
  group_by = NULL,
  split_by = NULL,
  facet_by = NULL,
  xlab = NULL,
  ylab = NULL,
  x_text_angle = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the feature data and metadata.}

\item{features}{A character vector of feature names}

\item{plot_type}{Type of the plot. It can be "violin", "box", "bar", "ridge", "dim", "cor", "heatmap" or "dot"}

\item{graph}{Specify the graph name to add edges between cell neighbors to the plot, only used when \code{plot_type} is "dim".}

\item{bg_cutoff}{Background cutoff for the dim plot, only used when \code{plot_type} is "dim".}

\item{dims}{Dimensions to plot, only used when \code{plot_type} is "dim".}

\item{rows_name}{The name of the rows in the heatmap, only used when \code{plot_type} is "heatmap".}

\item{ident}{The column name in the meta data to identify the cells.}

\item{agg}{The aggregation function to use for the bar plot.}

\item{group_by}{The column name in the meta data to group the cells.}

\item{split_by}{Column name in the meta data to split the cells to different plots.
If TRUE, the cells are split by the features.}

\item{facet_by}{Column name in the meta data to facet the plots. Should be always NULL.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{x_text_angle}{The angle of the x-axis text. Only used when \code{plot_type} is "violin", "bar", or "box".}

\item{...}{
  Arguments passed on to \code{\link[=FeatureStatPlot]{FeatureStatPlot}}
  \describe{
    \item{\code{object}}{A seurat object}
    \item{\code{spat_unit}}{The spatial unit to use for the plot. Only applied to Giotto objects.}
    \item{\code{feat_type}}{feature type of the features (e.g. "rna", "dna", "protein"), only applied to Giotto objects.}
    \item{\code{reduction}}{Name of the reduction to plot (for example, "umap"), only used when \code{plot_type} is "dim" or you can to use the reduction as feature.}
    \item{\code{assay}}{The assay to use for the feature data.}
    \item{\code{layer}}{The layer to use for the feature data.}
  }}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Feature statistic plot when given a composed data frame
}
\keyword{internal}
