% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurestatplot.R
\name{.feature_stat_plot}
\alias{.feature_stat_plot}
\title{Feature statistic plot when given a composed data frame}
\usage{
.feature_stat_plot(
  data,
  features,
  plot_type,
  should_shrink,
  should_pivot,
  downsample,
  graph = NULL,
  bg_cutoff = 0,
  dims = 1:2,
  rows_name = "Features",
  ident = NULL,
  agg = mean,
  group_by = NULL,
  pos_only = c("no", "any", "all"),
  split_by = NULL,
  facet_by = NULL,
  xlab = NULL,
  ylab = NULL,
  x_text_angle = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the feature data and metadata.}

\item{features}{A character vector of feature names}

\item{plot_type}{Type of the plot. It can be "violin", "box", "bar", "ridge", "dim", "cor", "heatmap" or "dot"}

\item{downsample}{A numeric the number of cells in each identity group to downsample to for violin, box, or ridge plots.
If n > 1, it is treated as the number of cells to downsample to.
If 0 < n <= 1, it is treated as the fraction of cells to downsample to.}

\item{graph}{Specify the graph name to add edges between cell neighbors to the plot, only used when \code{plot_type} is "dim".}

\item{bg_cutoff}{Background cutoff for the dim plot, only used when \code{plot_type} is "dim".}

\item{dims}{Dimensions to plot, only used when \code{plot_type} is "dim".}

\item{rows_name}{The name of the rows in the heatmap, only used when \code{plot_type} is "heatmap".}

\item{ident}{The column name in the meta data to identify the cells.}

\item{agg}{The aggregation function to use for the bar plot.}

\item{group_by}{The column name in the meta data to group the cells.}

\item{pos_only}{Whether to only include cells with positive feature values.
\itemize{
\item "no": Do not filter cells based on feature values. (default)
\item "any": Include cells with positive values for any of the features.
\item "all": Include cells with positive values for all of the features.
If you have named features (i.e. a named list), \code{pos_only} will be applied to all flattened features.
}}

\item{split_by}{Column name in the meta data to split the cells to different plots.
If TRUE, the cells are split by the features.}

\item{facet_by}{Column name in the meta data to facet the plots. Should be always NULL.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{x_text_angle}{The angle of the x-axis text. Only used when \code{plot_type} is "violin", "bar", or "box".}

\item{...}{
  Arguments passed on to \code{\link[=FeatureStatPlot]{FeatureStatPlot}}
  \describe{
    \item{\code{object}}{A seurat object, a giotto object, a path to an .h5ad file or an opened \code{H5File} by \code{hdf5r} package.}
    \item{\code{spat_unit}}{The spatial unit to use for the plot. Only applied to Giotto objects.}
    \item{\code{feat_type}}{feature type of the features (e.g. "rna", "dna", "protein"), only applied to Giotto objects.}
    \item{\code{reduction}}{Name of the reduction to plot (for example, "umap"), only used when \code{plot_type} is "dim" or you can to use the reduction as feature.}
    \item{\code{assay}}{The assay to use for the feature data.}
    \item{\code{layer}}{The layer to use for the feature data.}
  }}
}
\value{
A ggplot object or a list if \code{combine} is FALSE
}
\description{
Feature statistic plot when given a composed data frame
}
\keyword{internal}
