% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialutils.R
\name{.seurat_points_layer}
\alias{.seurat_points_layer}
\title{Process points layer for Seurat spatial plots}
\usage{
.seurat_points_layer(
  object,
  fov = NULL,
  boundaries = NULL,
  x = "x",
  y = "y",
  swap_xy = TRUE,
  image,
  args,
  crop,
  points_data,
  ext_unscaled,
  scale_factor,
  group_by,
  features,
  layer,
  legend.position,
  legend.direction,
  flip_y,
  ext
)
}
\arguments{
\item{object}{A Seurat object or a Giotto object.}

\item{fov}{The name of the field of view (FOV) to plot, only works for Seurat objects.}

\item{boundaries}{The name of the boundaries to plot, only works for Seurat objects.}

\item{x}{Internal use only, the name of the x coordinate column in the data. Used to adopt different data types.}

\item{y}{Internal use only, the name of the y coordinate column in the data. Used to adopt different data types.}

\item{swap_xy}{Logical, whether to swap x and y coordinates in the points data.
Seurat objects, loaded with Visium, Xenium or SlideSeq data, have x and y coordinates
in the opposite order. But when loaded with FOV data, the coordinates are in the correct order.}

\item{image}{The name of the image to plot. Possible values are:
\itemize{
\item NULL: For Seurat objects with Visium data, the first image will be used.
For Giotto objects and Seurat objects with other spatial data, no image will be plotted.
\item image name(s): the name of the image(s) to plot.
\item color name: a color to use as a background for the plot.
\item TRUE: For Seurat objects with Visium data, the first image will be used.
For Seurat objects with other spatial data, an error will be raised.
For Giotto objects with FOV, all (non-overlapping) images will be plotted; otherwise first image will be used.
\item FALSE: disable image plotting.
}}

\item{crop}{Whether to crop the plot to the extent of the available data. Similar to \code{crop} argument in
\code{\link[Seurat:SpatialPlot]{Seurat::SpatialDimPlot()}}. Defaults to TRUE.}

\item{scale_factor}{Internal use only. The scale factor to use for the image, which will be extracted from the object.}

\item{group_by}{The name of the metadata column to group the points by. Should be a character or factor column.
A special value "molecules" can be used to plot molecules in the FOV.}

\item{features}{A character vector of feature names to plot. If provided, the points will be colored by the features.
For \code{SpatDimPlot()}, this will be used to plot the molecules in the FOV.
For \code{SpatFeaturePlot()}, the plots will be faceted by the features.}

\item{layer}{The layer to use for the feature expression data.
Applicable for both Seurat and Giotto objects.
Defaults to "data" for Seurat objects, and "normalized" for Giotto objects.
For Giotto objects, it can also be "scaled", "raw", "counts", or "custom".
For Seurat objects, it can be "data", "scale.data", or "counts".}

\item{legend.position}{The position of the legend. Defaults to "right".}

\item{legend.direction}{The direction of the legend. Defaults to "vertical".}

\item{flip_y}{Internal use mostly, unless you want to flip the y-axis of the plot.}

\item{ext}{The extent of the plot. If NULL, the extent will be calculated from the data.
If a numeric vector of length 4, it should be in the format c(xmin, xmax, ymin, ymax).
It can also be an object created by \code{\link[terra:ext]{terra::ext()}}.}
}
\value{
A list containing the ggplot2 layer object and the facet_by variable if applicable.
}
\description{
Process points layer for Seurat spatial plots
}
\keyword{internal}
