% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentplot.R
\name{EnrichmentPlot}
\alias{EnrichmentPlot}
\title{Enrichment Plot}
\usage{
EnrichmentPlot(
  data,
  top_term = NULL,
  plot_type = c("bar", "dot", "lollipop", "network", "enrichmap", "wordcloud",
    "comparison"),
  x_by = NULL,
  size_by = NULL,
  fill_cutoff_name = NULL,
  fill_name = NULL,
  character_width = 50,
  expand = NULL,
  word_type = c("term", "feature"),
  split_by = NULL,
  split_by_sep = "_",
  facet_by = NULL,
  facet_scales = NULL,
  group_by = NULL,
  group_by_sep = "_",
  metric = "p.adjust",
  cutoff = NULL,
  palette = "Spectral",
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame with enrichment results generated by \code{clusterProfiler}.
If your result is generated by \code{enrichr}, you can use \link{PrepareEnrichrResult} to convert it to a data frame.}

\item{top_term}{Number of top terms to show in the plot.
Default is 6 for all plots except "enrichmap" which is 100.}

\item{plot_type}{Type of plot to generate. Options are "bar", "dot", "lollipop", "network", "enrichmap", "wordcloud", "comparison".}

\item{x_by}{A character vector of column names to use for the x-axis. Default is NULL.
Works only for "dot" and "lollipop".}

\item{size_by}{A character vector of column names to use for the size of the points. Default is NULL.
Works only for "comparison", "dot" and "lollipop".}

\item{fill_cutoff_name}{The legend name for the terms above the cutoff value. Default is NULL.
Works only for "comparison", "dot" and "lollipop".}

\item{fill_name}{The legend name for the metric. Default is NULL.
Works only for "comparison", "dot" and "lollipop".}

\item{character_width}{The width of the terms in the plot. Default is 50.
When the terms are too long, they will be wrapped to fit the width.}

\item{expand}{A numeric vector of length 1, 2 or 4 to expand the plot. Default is NULL.
Works only for "bar" plot.
See also \link[plotthis:barplot]{plotthis::BarPlot}.}

\item{word_type}{The type of word to show in the wordcloud. Options are "term" and "feature". Default is "term".
Works only for "wordcloud".}

\item{split_by}{A character vector of column names to split the plots. Default is NULL.}

\item{split_by_sep}{A character to separate the split_by column names. Default is "_".}

\item{facet_by}{A character vector of column names to facet the plots. Default is NULL.}

\item{facet_scales}{The facet scales. Default is NULL.}

\item{group_by}{A character vector of column names to group the terms. Default is NULL.
Works only for "comparison" plot.}

\item{group_by_sep}{A character to concatenate the group_by columns when there are multiple columns. Default is "_".
Works only for "comparison" plot.}

\item{metric}{The column name to use for the metric. Default is "p.adjust".}

\item{cutoff}{The cutoff value to the metric to mark the tems on the plot. Default is NULL.
Note that the terms are not filtered by this value. Use \code{top_terms} to filter the terms.
When specified:
\itemize{
\item 'bar' plot will show a line at the cutoff value
\item 'dot' plot will show gray dots for the terms above the cutoff value
\item 'lollipop' plot will show gray dots for the terms above the cutoff value
\item 'comparison' plot will show gray dots for the terms above the cutoff value
\item has no effect on 'network', 'enrichmap', and 'wordcloud' plots
}}

\item{palette}{The color palette to use for the plot. Default is "Spectral".
See \link[plotthis:show_palettes]{plotthis::show_palettes} for available palettes.}

\item{xlab}{The x-axis label. Default is NULL.}

\item{ylab}{The y-axis label. Default is NULL.}

\item{...}{Other arguments passed to the specific plot function.
\itemize{
\item For "bar", \code{\link[plotthis:barplot]{plotthis::BarPlot()}}.
\item For "dot", \code{\link[plotthis:dotplot]{plotthis::DotPlot()}}.
\item For "lollipop", \code{\link[plotthis:dotplot]{plotthis::LollipopPlot()}}.
\item For "network", \code{\link[plotthis:enrichmap1]{plotthis::EnrichNetwork()}}.
\item For "enrichmap", \code{\link[plotthis:enrichmap1]{plotthis::EnrichMap()}}.
\item For "wordcloud", \code{\link[plotthis:WordCloudPlot]{plotthis::WordCloudPlot()}}.
\item For "comparison", \code{\link[plotthis:dotplot]{plotthis::DotPlot()}}.
}}
}
\description{
This function generates various types of plots for enrichment (over-representation) analysis.
}
\examples{
set.seed(8525)
data(enrich_example, package = "plotthis")
data(enrich_multidb_example, package = "plotthis")

EnrichmentPlot(enrich_example)
EnrichmentPlot(enrich_example, cutoff = 0.05)
EnrichmentPlot(enrich_example, palette = "Paired")

# Multiple databases#'
EnrichmentPlot(enrich_multidb_example, facet_by = "Database", facet_nrow = 2)

enrich_example$Group <- sample(c("A", "B"), nrow(enrich_example), replace = TRUE)
EnrichmentPlot(enrich_example, plot_type = "comparison", group_by = "Group")
EnrichmentPlot(enrich_example, plot_type = "dot", top_term = 10)
EnrichmentPlot(enrich_example, plot_type = "lollipop", top_term = 10)
EnrichmentPlot(enrich_example, plot_type = "network")
EnrichmentPlot(enrich_example, plot_type = "enrichmap")
EnrichmentPlot(enrich_example, plot_type = "wordcloud")
# Wordcloud with feature
EnrichmentPlot(enrich_example, plot_type = "wordcloud", word_type = "feature")
}
