% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalutils.R
\name{get_clonal_grouping_levels}
\alias{get_clonal_grouping_levels}
\title{Get the grouping levels for clonal data for later to restore}
\usage{
get_clonal_grouping_levels(data, groupings, order = NULL)
}
\arguments{
\item{data}{The product of \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, \link[scRepertoire:combineTCR]{scRepertoire::combineTCR}, or \link[scRepertoire:combineExpression]{scRepertoire::combineExpression}.}

\item{groupings}{The column names in the meta data to group the cells.}

\item{order}{A list specifying the order of the levels for each grouping variable.}
}
\value{
A list of the grouping levels for each grouping variable.
}
\description{
Data transform with multiple groupings may lose the original grouping levels.
This function is to get the grouping levels for later to restore.
}
\keyword{internal}
