% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markersplot.R
\name{MarkersPlot}
\alias{MarkersPlot}
\title{Visualize Markers}
\usage{
MarkersPlot(
  markers,
  object = NULL,
  plot_type = c("volcano", "volcano_log2fc", "volcano_pct", "jitter", "jitter_log2fc",
    "jitter_pct", "heatmap_log2fc", "heatmap_pct", "dot_log2fc", "dot_pct", "heatmap",
    "violin", "box", "bar", "ridge", "dot"),
  subset_by = NULL,
  subset_as_facet = FALSE,
  comparison_by = NULL,
  p_adjust = TRUE,
  cutoff = NULL,
  order_by = NULL,
  select = ifelse(plot_type \%in\% c("volcano", "volcano_log2fc", "volcano_pct",
    "jitter", "jitter_log2fc", "jitter_pct"), 5, 10),
  ...
)
}
\arguments{
\item{markers}{A data frame of markers, typically identified by \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} or \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}.}

\item{object}{A Seurat object. Required for some plot types, see \code{plot_type}.}

\item{plot_type}{Type of plot to generate. Options include:
\itemize{
\item \code{volcano}/\code{volcano_log2fc}: Volcano plot with log2 fold change on x-axis and -log10(p-value) on y-axis.
If \code{p_adjust} is TRUE, -log10(adjusted p-value) is used instead.
\item \code{volcano_pct}: Volcano plot with difference in percentage of cells expressing the gene between two groups on x-axis and -log10(p-value) on y-axis.
If \code{p_adjust} is TRUE, -log10(adjusted p-value) is used instead.
\item \code{jitter}/\code{jitter_log2fc}: Jitter plot of log2 fold change for each gene.
The x-axis is the groups defined by \code{subset_by}, and the y-axis is log2 fold change.
The size of the dots represents -log10(p-value) or -log10(adjusted p-value) if \code{p_adjust} is TRUE.
\item \code{jitter_pct}: Jitter plot of difference in percentage of cells expressing the gene between two groups for each gene.
The x-axis is the groups defined by \code{subset_by}, and the y-axis is the difference in percentage of cells expressing the gene between two groups.
The size of the dots represents -log10(p-value) or -log10(adjusted p-value) if \code{p_adjust} is TRUE.
\item \code{heatmap_log2fc}: Heatmap of log2 fold change for each gene across groups defined by \code{subset_by}. By specifying \code{cutoff},
The heatmap cells will be labeled with "*" for p-value < cutoff, if \code{p_adjust} is TRUE, adjusted p-value < cutoff.
\item \code{heatmap_pct}: Heatmap of difference in percentage of cells expressing the gene between two groups for each gene across groups defined by \code{subset_by}.
By specifying \code{cutoff}, The heatmap cells will be labeled with "*" for p-value < cutoff, if \code{p_adjust} is TRUE, adjusted p-value < cutoff.
\item \code{dot_log2fc}: Dot plot of log2 fold change for each gene across groups defined by \code{subset_by}.
The size of the dots represents -log10(p-value) or -log10(adjusted p-value) if \code{p_adjust} is TRUE.
\item \code{dot_pct}: Dot plot of difference in percentage of cells expressing the gene between two groups for each gene across groups defined by \code{subset_by}.
The size of the dots represents -log10(p-value) or -log10(adjusted p-value) if \code{p_adjust} is TRUE.
\item \code{heatmap}: Heatmap of expression values for each gene across groups defined by \code{subset_by}. Requires \code{object}.
\item \code{violin}: Violin plot of expression values for each gene across groups defined by \code{subset_by}. Requires \code{object}.
\item \code{box}: Box plot of expression values for each gene across groups defined by \code{subset_by}. Requires \code{object}.
\item \code{bar}: Bar plot of average expression values for each gene across groups defined by \code{subset_by}. Requires \code{object}.
\item \code{ridge}: Ridge plot of expression values for each gene across groups defined by \code{subset_by}. Requires \code{object}.
\item \code{dot}: Dot plot of expression values for each gene across groups defined by \code{subset_by}. Requires \code{object}.
}}

\item{subset_by}{A column in markers indicating where the markers are identified from, e.g., cluster or condition.
If object is provided, you can provide the corresponding metadata column in \code{subset_by} to merge the markers with the object's metadata, using \code{:} as separator. For example, if the markers are identified from different clusters (identified by \code{FindAllMarkers}), and the object's ident column is "RNA_snn_res.0.8", you can set \code{subset_by = "cluster:RNA_snn_res.0.8"}.
Note that for other columns to be merged, only the first value of each group defined by \code{subset_by} will be used.
For some plots, this is used to split the markers into multiple plots, e.g., one plot for each cluster.
See \code{plot_type} for details.}

\item{subset_as_facet}{Logical, whether to facet the plots by \code{subset_by} if applicable.}

\item{comparison_by}{The metadata column in markers indicating the comparion.
When visualizing the expression values, this column should also be in the object's metadata.
The values of this column should either a single value, indicating the comparison is between this group and all other cells (in the subset),
Similar as \code{subset_by}, you can provide the corresponding object metadata column by using \code{:} as separator.
If \code{NULL}, all markers are treated as from one comparison.}

\item{p_adjust}{Logical, whether to use adjusted p-value for plots that involve p-values.
Default is TRUE.}

\item{cutoff}{Numeric, p-value or adjusted p-value cutoff to label significance in heatmap plots.
Default is NULL, no cutoff.}

\item{order_by}{A string of expression to order the markers within each group defined by \code{subset_by}.
In addition to the columns in \code{markers}, you can also use the columns from the object's metadata if \code{object} is provided.
The object's metadata will be merged with \code{markers} by \code{subset_by}. Be carefull that only the first value of other columns will be used.}

\item{select}{Number of top markers (ordered by \code{order_by}) to select for each group defined by \code{subset_by} or a string of expression to filter markers.
It will be evaluated by \code{\link[dplyr:filter]{dplyr::filter()}}.
For \code{volcano}, \code{volcano_log2fc}, \code{volcano_pct}, \code{jitter}, \code{jitter_log2fc}, and \code{jitter_pct} plots, the selected markers will be labeled in the plot.
FOr  other plot types, only the selected markers will be plotted.
Default is 5 for \code{volcano}, \code{volcano_log2fc}, \code{volcano_pct}, \code{jitter}, \code{jitter_log2fc}, \code{jitter_pct}, and 10 for other plot types.}

\item{...}{Additional arguments passed to specific plotting functions.
See Details.}
}
\value{
A ggplot object or a list of ggplot objects if not merged.
}
\description{
Plot markers, typically identified by \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} or \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}.
}
\details{
Additional arguments passed to specific plotting functions:
\itemize{
\item For \code{heatmap_log2fc}, \code{heatmap_pct}, \code{dot_log2fc}, and \code{dot_pct}, the arguments will be passed to \code{\link[plotthis:Heatmap]{plotthis::Heatmap()}}
\item For \code{violin}, "box", "bar", "ridge" and "dot", the arguments will be passed to \code{\link[=FeatureStatPlot]{FeatureStatPlot()}}
\item For \code{volcano}, \code{volcano_log2fc}, \code{volcano_pct}, the arguments will be passed to \code{\link[plotthis:VolcanoPlot]{plotthis::VolcanoPlot()}}
\item For \code{jitter}, \code{jitter_log2fc}, \code{jitter_pct}, the arguments will be passed to \code{\link[plotthis:jitterplot]{plotthis::JitterPlot()}}
}
}
\examples{
\donttest{
data(pancreas_sub)
markers <- Seurat::FindMarkers(pancreas_sub,
 group.by = "Phase", ident.1 = "G2M", ident.2 = "G1")
allmarkers <- Seurat::FindAllMarkers(pancreas_sub)  # seurat_clusters

MarkersPlot(markers)
MarkersPlot(markers, x_cutoff = 2)
MarkersPlot(allmarkers,
    subset_by = "cluster", ncol = 2, subset_as_facet = TRUE)
MarkersPlot(markers, plot_type = "volcano_pct", flip_negative = TRUE)

MarkersPlot(allmarkers, plot_type = "jitter", subset_by = "cluster")
MarkersPlot(allmarkers, plot_type = "jitter_pct",
    subset_by = "cluster", add_hline = 0, shape = 16)

MarkersPlot(allmarkers, plot_type = "heatmap_log2fc", subset_by = "cluster")
MarkersPlot(allmarkers, plot_type = "heatmap_pct", subset_by = "cluster",
    cutoff = 0.05)

MarkersPlot(allmarkers, plot_type = "dot_log2fc", subset_by = "cluster",
    add_reticle = TRUE)

MarkersPlot(allmarkers, object = pancreas_sub, plot_type = "heatmap",
   columns_split_by = "CellType",
   comparison_by = "cluster:seurat_clusters")

# Suppose we did a DE between g1 and g2 in each cluster
allmarkers$comparison <- "g1:g2"
MarkersPlot(allmarkers, object = pancreas_sub, plot_type = "heatmap",
   comparison_by = "Phase", subset_by = "cluster:seurat_clusters")
MarkersPlot(allmarkers, object = pancreas_sub, plot_type = "dot",
   comparison_by = "Phase", subset_by = "cluster:seurat_clusters")

MarkersPlot(allmarkers, object = pancreas_sub, plot_type = "violin", select = 3,
   comparison_by = "Phase", subset_by = "cluster:seurat_clusters")

MarkersPlot(allmarkers, object = pancreas_sub, plot_type = "box", select = 3,
   comparison_by = "Phase", subset_by = "cluster:seurat_clusters")

MarkersPlot(allmarkers, object = pancreas_sub, plot_type = "ridge", select = 2,
   comparison_by = "Phase", subset_by = "cluster:seurat_clusters",
   ncol = 2)
}
}
