% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{SCPlotterChat}
\alias{SCPlotterChat}
\title{SCPlotterChat Class}
\description{
SCPlotterChat Class

SCPlotterChat Class
}
\details{
An R6 class that provides chat functionality for SCPlotter
}
\examples{
\donttest{
if (FALSE) {
provider <- tidyprompt::llm_provider_openai(api_key = Sys.getenv("OPENAI_API_KEY"))
chat <- SCPlotterChat$new(provider)
chat$ask("What are the tools to use?")
# Tool identified:  ListTools
# Available tools:
# -  ClonalOverlapPlot :  ClonalOverlapPlot
#    Plot the overlap of the clones in different samples/groups.
#
# -  EnrichmentPlot :  Enrichment Plot
#    This function generates various types of plots for enrichment analysis.
# ...

chat$ask("Plot the default cell-cell communication plot for the cellphonedb_res dataset")
# Tool identified:  CCCPlot
# Data object identified:  cellphonedb_res
# Running tool:  CCCPlot

chat$ask("do a heatmap instead")
# Tool identified:  CCCPlot
# Data object identified:  cellphonedb_res
# Running tool:  CCCPlot
}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SCPlotterChat-new}{\code{SCPlotterChat$new()}}
\item \href{#method-SCPlotterChat-clear_history}{\code{SCPlotterChat$clear_history()}}
\item \href{#method-SCPlotterChat-get_history}{\code{SCPlotterChat$get_history()}}
\item \href{#method-SCPlotterChat-list_tools}{\code{SCPlotterChat$list_tools()}}
\item \href{#method-SCPlotterChat-list_data}{\code{SCPlotterChat$list_data()}}
\item \href{#method-SCPlotterChat-set_data}{\code{SCPlotterChat$set_data()}}
\item \href{#method-SCPlotterChat-get_data}{\code{SCPlotterChat$get_data()}}
\item \href{#method-SCPlotterChat-ask}{\code{SCPlotterChat$ask()}}
\item \href{#method-SCPlotterChat-clone}{\code{SCPlotterChat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-new"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of the SCPlotterChat class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$new(provider, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider}}{An LLM provider object
See https://tjarkvandemerwe.github.io/tidyprompt/articles/getting_started.html#setup-an-llm-provider
for more details on LLM providers}

\item{\code{verbose}}{A logical value indicating whether to print verbose messages
Default is FALSE. And this will override the verbose setting of the provider}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the SCPlotterChat class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-clear_history"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-clear_history}{}}}
\subsection{Method \code{clear_history()}}{
Clear the chat history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$clear_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-get_history}{}}}
\subsection{Method \code{get_history()}}{
Get the chat history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$get_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The chat history
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-list_tools"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-list_tools}{}}}
\subsection{Method \code{list_tools()}}{
Print the list of available tools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$list_tools()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-list_data"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-list_data}{}}}
\subsection{Method \code{list_data()}}{
Print the list of available data objects that can be used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$list_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-set_data}{}}}
\subsection{Method \code{set_data()}}{
Set the data to be analyzed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$set_data(data, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data object to be set}

\item{\code{name}}{The name of the data object (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data to be analyzed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The data object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-ask"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-ask}{}}}
\subsection{Method \code{ask()}}{
Send a prompt to the chat interface and receive a response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$ask(prompt, verbose = NULL, add_to_history = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prompt}}{A character string containing the user's query or instruction}

\item{\code{verbose}}{A logical value indicating whether to print verbose messages
Default is NULL, which will use the verbose setting of the SCPlotterChat object}

\item{\code{add_to_history}}{A logical value indicating whether to add the prompt and response to the chat history
Default is TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A response from the chat system
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCPlotterChat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SCPlotterChat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPlotterChat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
